% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.drclist.R
\name{anova.drclist}
\alias{anova.drclist}
\title{ANOVA for list of drc objects}
\usage{
\method{anova}{drclist}(object, ..., details = TRUE, test = NULL)
}
\description{
ANOVA for list of drc objects
}
\keyword{internal}
