% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{AR.2}
\alias{AR.2}
\title{Two-parameter asymptotic regression model}
\usage{
AR.2(fixed = c(NA, NA), names = c("d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 2, specifying fixed parameters (use \code{NA} for
parameters that should be estimated).}

\item{names}{character vector of length 2 giving the names of the parameters
(default \code{c("d", "e")}).}

\item{...}{additional arguments passed to \code{\link{weibull2}}, most
notably \code{method} (a character string: \code{"1"} (default),
\code{"2"}, \code{"3"}, or \code{"4"}) which selects the self-starter
method for obtaining starting values. See \code{\link{weibull2}} for
details.}
}
\value{
A list of class \code{"Weibull-2"} as returned by \code{\link{weibull2}}.
}
\description{
A two-parameter asymptotic regression model where \code{b} is fixed at 1 and
the lower limit is fixed at 0. The model is given by the equation
\deqn{f(x) = d \cdot (1 - \exp(-x / e))}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = AR.2())

}
\seealso{
\code{\link{AR.3}}, \code{\link{weibull2}}, \code{\link{EXD.2}}
}
\keyword{models}
\keyword{nonlinear}
