% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.R
\name{BC.4}
\alias{BC.4}
\title{Four-parameter Brain-Cousens hormesis model}
\usage{
BC.4(fixed = c(NA, NA, NA, NA), names = c("b", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4 specifying fixed parameters (NAs for free parameters).}

\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{braincousens}}.}
}
\value{
A list (see \code{\link{braincousens}}).
}
\description{
\code{BC.4} provides the Brain-Cousens modified log-logistic model with the lower limit fixed at 0.
}
\examples{
lettuce.bcm2 <- drm(weight ~ conc, data = lettuce, fct = BC.4())
summary(lettuce.bcm2)
ED(lettuce.bcm2, c(50))

}
\references{
van Ewijk, P. H. and Hoekstra, J. A. (1993)
Calculation of the EC50 and its Confidence Interval When Subtoxic Stimulus Is Present,
\emph{Ecotoxicology and Environmental Safety}, \bold{25}, 25--32.
}
\seealso{
\code{\link{braincousens}}, \code{\link{BC.5}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
