% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.R
\name{BC.5}
\alias{BC.5}
\title{Five-parameter Brain-Cousens hormesis model}
\usage{
BC.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 5 specifying fixed parameters (NAs for free parameters).}

\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{braincousens}}.}
}
\value{
A list (see \code{\link{braincousens}}).
}
\description{
\code{BC.5} provides the full five-parameter Brain-Cousens modified log-logistic model
for describing hormesis.
}
\examples{
lettuce.bcm1 <- drm(weight ~ conc, data = lettuce, fct = BC.5())
modelFit(lettuce.bcm1)
plot(lettuce.bcm1)

}
\seealso{
\code{\link{braincousens}}, \code{\link{BC.4}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
