% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.R
\name{bcl4}
\alias{bcl4}
\title{Alias for BC.5}
\usage{
bcl4(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 5 specifying fixed parameters (NAs for free parameters).}

\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{braincousens}}.}
}
\description{
\code{bcl4} is an alias for \code{\link{BC.5}}.
}
\seealso{
\code{\link{BC.5}}
}
\keyword{models}
\keyword{nonlinear}
