\name{blackgrass}
\alias{blackgrass}
\docType{data}
\title{Seedling Emergence of Blackgrass (Alopecurus myosuroides)}
\description{Seedling emergence of herbicide susceptible (S) and resistant (R) Alopecurus myosuroides  in reponse to sowing depth and suboptimal temperature regimes (10/5C) and optimal temperature regimes (17/10C).}

\usage{data("blackgrass")}

\format{
  A data frame with 2752 observations on the following 12 variables.
  \describe{
    \item{\code{Exp}}{a numeric vector}
    \item{\code{Temp}}{a numeric vector}
    \item{\code{Popu}}{a numeric vector}
    \item{\code{Bio}}{a factor with two levels}
    \item{\code{Depth}}{a numeric vector}
    \item{\code{Rep}}{a numeric vector}
    \item{\code{Start.Day}}{a numeric vector}
    \item{\code{End.Day}}{a numeric vector}
    \item{\code{Ger}}{a numeric vector}
    \item{\code{Accum.Ger}}{a numeric vector}
    \item{\code{TotalSeed}}{a numeric vector}
    \item{\code{Pot}}{a numeric vector}
  }
}
\references{Keshtkar, E., Mathiassen, S. K., Beffa, R., Kudsk, P. (2017). Seed Germination and Seedling Emergence of Blackgrass (Alopecurus myosuroides) as Affected by Non-Target-Site Herbicide Resistance. Weed Science, 65, 732-742. https://doi.org/10.1017/wsc.2017.44
%%  ~~ possibly secondary sources and usages ~~
}

\examples{
library(drc)

## Displaying the data
head(blackgrass)

## Summarizing seedling emergence across treatments
aggregate(Accum.Ger ~ Temp + Bio, data = blackgrass, FUN = max)
}
\keyword{datasets}
