% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.R
\name{braincousens}
\alias{braincousens}
\title{The Brain-Cousens hormesis models}
\usage{
braincousens(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e, f.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list containing the non-linear function, the self starter function,
the parameter names and additional model specific objects.
}
\description{
\code{braincousens} provides a very general way of specifying Brain-Cousens'
modified log-logistic model for describing hormesis, under various constraints on the parameters.
}
\details{
The Brain-Cousens model is given by the expression
\deqn{f(x) = c + \frac{d-c+fx}{1+\exp(b(\log(x)-\log(e)))}}
which is a five-parameter model.
}
\references{
Brain, P. and Cousens, R. (1989) An equation to describe dose responses
where there is stimulation of growth at low doses,
\emph{Weed Research}, \bold{29}, 93--96.
}
\seealso{
\code{\link{BC.4}}, \code{\link{BC.5}}, \code{\link{drm}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
