% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.ssf.R
\name{braincousens.ssf}
\alias{braincousens.ssf}
\title{Self-starter for Brain-Cousens model}
\usage{
braincousens.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for Brain-Cousens model
}
\keyword{internal}
