% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcompX.R
\name{CIcomp}
\alias{CIcomp}
\title{Classical combination index for effective doses}
\usage{
CIcomp(mixProp, modelList, EDvec)
}
\arguments{
\item{mixProp}{a numeric value between 0 and 1 specifying the mixture proportion/ratio.}

\item{modelList}{a list containing 3 model fits using \code{\link{drm}}: the mixture model fit
first, followed by the 2 pure substance model fits.}

\item{EDvec}{a numeric vector of effect levels (percentages between 0 and 100).}
}
\value{
A matrix with one row per ED value. Columns contain estimated combination indices,
  their standard errors and 95\% confidence intervals, p-value for testing CI=1, estimated
  ED values for the mixture data and assuming concentration addition (CA) with corresponding
  standard errors.
}
\description{
Calculates the classical combination index for effective doses in binary mixture experiments.
}
\examples{
## Fitting marginal models for the 2 pure substances
acidiq.0 <- drm(rgr ~ dose, data = subset(acidiq, pct == 999 | pct == 0), fct = LL.4())
acidiq.100 <- drm(rgr ~ dose, data = subset(acidiq, pct == 999 | pct == 100), fct = LL.4())

## Fitting model for single mixture with ratio 17:83
acidiq.17 <- drm(rgr ~ dose, data = subset(acidiq, pct == 17 | pct == 0), fct = LL.4())

## Calculation of combination indices based on ED10, ED20, ED50
CIcomp(0.17, list(acidiq.17, acidiq.0, acidiq.100), c(10, 20, 50))

}
\references{
Martin-Betancor, K. and Ritz, C. and Fernandez-Pinas, F. and Leganes, F. and
  Rodea-Palomares, I. (2015) Defining an additivity framework for mixture research in
  inducible whole-cell biosensors, \emph{Scientific Reports} \bold{17200}.
}
\seealso{
\code{\link{CIcompX}}, \code{\link{plotFACI}}, \code{\link{mixture}}
}
\author{
Christian Ritz and Ismael Rodea-Palomares
}
\concept{antagonism mixture synergy}
\keyword{models}
\keyword{nonlinear}
