% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcompX.R
\name{CIcompX}
\alias{CIcompX}
\title{Calculation of combination index for binary mixtures}
\usage{
CIcompX(mixProp, modelList, EDvec, EDonly = FALSE)
}
\arguments{
\item{mixProp}{a numeric value between 0 and 1 specifying the mixture proportion/ratio.}

\item{modelList}{a list containing 3 model fits using \code{\link{drm}}: the mixture model fit
first, followed by the 2 pure substance model fits.}

\item{EDvec}{a numeric vector of effect levels (percentages between 0 and 100).}

\item{EDonly}{logical. If TRUE, only combination indices for effective doses are calculated.}
}
\value{
A list with components \code{Effx}, \code{Effy} (unless \code{EDonly = TRUE}),
\code{CAx}, \code{CAy} (unless \code{EDonly = TRUE}), and \code{EDvec}.
}
\description{
For single mixture data, combination indices for effective doses as well as effects
may be calculated. This is an extended version of \code{\link{CIcomp}}.
}
\references{
Martin-Betancor, K. and Ritz, C. and Fernandez-Pinas, F. and Leganes, F. and
Rodea-Palomares, I. (2015) Defining an additivity framework for mixture research in
inducible whole-cell biosensors, \emph{Scientific Reports} \bold{17200}.
}
\seealso{
\code{\link{CIcomp}}, \code{\link{plotFACI}}, \code{\link{mixture}}
}
\author{
Christian Ritz and Ismael Rodea-Palomares
}
\concept{antagonism mixture synergy}
\keyword{models}
\keyword{nonlinear}
