% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks.distance.drc.R
\name{cooks.distance.drc}
\alias{cooks.distance.drc}
\title{Cook's distance for nonlinear dose-response models}
\usage{
\method{cooks.distance}{drc}(model, ...)
}
\arguments{
\item{model}{an object of class 'drc'.}

\item{...}{additional arguments (not used).}
}
\value{
A vector of Cook's distance values, one value per observation.
}
\description{
Cook's distance values are provided for nonlinear dose-response model fits using the
same formulas as in linear regression but based on the corresponding approximate quantities
available for nonlinear models.
}
\examples{
ryegrass.LL.4 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())

cooks.distance(ryegrass.LL.4)

}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
