% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDcomp.R
\name{createsifct}
\alias{createsifct}
\title{Create selectivity index function}
\usage{
createsifct(edfct, logBase = NULL, fls = FALSE, indexMat, lenCoef)
}
\description{
Create selectivity index function
}
\keyword{internal}
