% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRS.6.R
\name{CRS.6}
\alias{CRS.6}
\title{Generalised Cedergreen-Ritz-Streibig Model for Hormesis}
\usage{
CRS.6(
  fixed = c(NA, NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f", "g"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value
they are fixed. NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters
(should not contain ":").}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used (optional).}
}
\value{
A list containing the nonlinear model function, the self starter function,
  and the parameter names.
}
\description{
A six-parameter extension of the Cedergreen-Ritz-Streibig model for
describing hormesis, where the alpha parameter is estimated rather than fixed.
}
\details{
The model function is:

\deqn{f(x) = c + \frac{d-c+f \exp(-1/x^g)}{1+\exp(b(\log(x)-\log(e)))}}

This generalises the five-parameter \code{\link{CRS.5a}} model by estimating
the alpha exponent (parameter \eqn{g}) instead of fixing it.
}
\note{
This function is for use with \code{\link{drm}}.
}
\seealso{
\code{\link{CRS.5a}}, \code{\link{cedergreen}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
