% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmConvertParm.R
\name{drmConvertParm}
\alias{drmConvertParm}
\title{Convert parameter vectors to matrices}
\usage{
drmConvertParm(startVec, startMat, factor1, colList)
}
\description{
Convert parameter vectors to matrices
}
\keyword{internal}
