% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmEMbinomial.R
\name{drmEMbinomial}
\alias{drmEMbinomial}
\title{EM algorithm for binomial response}
\usage{
drmEMbinomial(
  dose,
  resp,
  multCurves,
  startVec,
  robustFct,
  weights,
  rmNA,
  zeroTol = 1e-12,
  doseScaling = 1,
  respScaling = 1
)
}
\description{
EM algorithm for binomial response
}
\keyword{internal}
