% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmEMls.R
\name{drmEMls}
\alias{drmEMls}
\title{EM algorithm for least squares}
\usage{
drmEMls(
  dose,
  resp,
  multCurves,
  startVec,
  robustFct,
  weights,
  rmNA,
  dmf = NULL,
  doseScaling = 1,
  respScaling = 1,
  varcov = NULL
)
}
\description{
EM algorithm for least squares
}
\keyword{internal}
