% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmOpt.R
\name{drmOpt}
\alias{drmOpt}
\title{Optimisation wrapper for drm}
\usage{
drmOpt(
  opfct,
  opdfct1,
  startVec,
  optMethod,
  constrained,
  warnVal,
  upperLimits,
  lowerLimits,
  errorMessage,
  maxIt,
  relTol,
  opdfct2 = NULL,
  parmVec,
  traceVal,
  silentVal = TRUE,
  matchCall
)
}
\description{
Optimisation wrapper for drm
}
\keyword{internal}
