% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmParNames.R
\name{drmParNames}
\alias{drmParNames}
\title{Generate parameter names for drm}
\usage{
drmParNames(
  numNames,
  parNames,
  collapseList2,
  repStr1 = "factor(pmodels[, i])",
  repStr2 = "factor(assayNo)"
)
}
\description{
Generate parameter names for drm
}
\keyword{internal}
