% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED.drc.R
\name{ED}
\alias{ED}
\title{Estimating effective doses}
\usage{
ED(object, ...)
}
\arguments{
\item{object}{an object of class \code{drc}.}

\item{...}{additional arguments passed to the method.}
}
\value{
See \code{\link{ED.drc}} for details on the return value.
}
\description{
S3 generic function that dispatches to the appropriate method for estimating
effective concentrations (EC) or effective doses (ED) at specified response
levels. For objects of class \code{drc}, the default method
\code{\link{ED.drc}} is called.
}
\seealso{
\code{\link{ED.drc}} for the default method, \code{\link{EDcomp}} for estimating differences and ratios of ED
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
