% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDhelper.R
\name{EDhelper}
\alias{EDhelper}
\title{Helper function for ED calculations}
\usage{
EDhelper(parmVec, respl, reference, typeCalc, cond = TRUE)
}
\description{
Helper function for ED calculations
}
\keyword{internal}
