% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDinvreg.R
\name{EDinvreg}
\alias{EDinvreg}
\title{Inverse regression for ED estimation}
\usage{
EDinvreg(
  object,
  respLev,
  catLev = NA,
  intType = "confidence",
  level,
  type,
  extFactor = 10
)
}
\description{
Inverse regression for ED estimation
}
\keyword{internal}
