% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{estfun.drc}
\alias{estfun.drc}
\title{Estimating function for the sandwich estimator}
\usage{
\method{estfun}{drc}(x, ...)
}
\arguments{
\item{x}{object of class \code{drc}.}

\item{...}{additional arguments. At the moment none are supported.}
}
\value{
The estimating function evaluated at the data and the parameter estimates.
By default no clustering is assumed, corresponding to robust standard errors
under independence.
}
\description{
Evaluates the estimating function (the "meat") for the sandwich estimator of the
variance-covariance matrix for objects of class 'drc'.
}
\details{
The details are provided by Zeileis (2006).
}
\examples{
## The lines below requires that the packages
## 'lmtest' and 'sandwich' are installed
# library(lmtest)
# library(sandwich)

# ryegrass.m1<-drm(rootl ~ conc, data = ryegrass, fct = LL.4())

# Standard summary output
# coeftest(ryegrass.m1)

# Output with robust standard errors
# coeftest(ryegrass.m1, vcov = sandwich)

}
\references{
Zeileis, A. (2006) Object-oriented Computation of Sandwich Estimators,
\emph{J. Statist. Software}, \bold{16}, Issue 9.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
