% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.drc.R
\name{fitted.drc}
\alias{fitted.drc}
\title{Extract fitted values from model}
\usage{
\method{fitted}{drc}(object, ...)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{...}{additional arguments.}
}
\value{
Fitted values extracted from \code{object}.
}
\description{
Extracts fitted values from an object of class 'drc'.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())
plot(fitted(ryegrass.m1), residuals(ryegrass.m1))  # a residual plot

}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
