% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplogistic.R
\name{FPL.4}
\alias{FPL.4}
\title{Four-parameter fractional polynomial-logistic model}
\usage{
FPL.4(p1, p2, fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{p1}{numeric denoting the negative power of log(dose+1) in the fractional polynomial.}

\item{p2}{numeric denoting the positive power of log(dose+1) in the fractional polynomial.}

\item{fixed}{numeric vector of length 4 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{fplogistic}}.}
}
\value{
A list (see \code{\link{fplogistic}}).
}
\description{
Convenience function for the four-parameter fractional polynomial-logistic model.
}
\seealso{
\code{\link{fplogistic}}, \code{\link{maED}}
}
\keyword{models}
\keyword{nonlinear}
