% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplogistic.R
\name{fplogistic}
\alias{fplogistic}
\title{Fractional polynomial-logistic dose-response model}
\usage{
fplogistic(
  p1,
  p2,
  fixed = c(NA, NA, NA, NA),
  names = c("b", "c", "d", "e"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{p1}{numeric denoting the negative power of log(dose+1) in the fractional polynomial.}

\item{p2}{numeric denoting the positive power of log(dose+1) in the fractional polynomial.}

\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list containing the nonlinear function, the self starter function
and the parameter names.
}
\description{
Model function for specifying dose-response models that are a combination of a logistic model
and an appropriate class of fractional polynomials.
}
\details{
The fractional polynomial dose-response models introduced by Namata et al. (2008) are implemented
using the logistic model as base.
}
\references{
Namata, Harriet and Aerts, Marc and Faes, Christel and Teunis, Peter (2008)
Model Averaging in Microbial Risk Assessment Using Fractional Polynomials,
\emph{Risk Analysis} \bold{28}, 891--905.
}
\seealso{
\code{\link{FPL.4}}, \code{\link{maED}}, \code{\link{drm}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
