% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{G.2}
\alias{G.2}
\title{Two-parameter Gompertz model}
\usage{
G.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)
}
\arguments{
\item{upper}{numeric specifying the fixed upper horizontal asymptote. Default is 1.}

\item{fixed}{numeric vector of length 2 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{gompertz}}.}
}
\value{
A list (see \code{\link{gompertz}}).
}
\description{
Convenience function for the Gompertz model with lower limit fixed at 0 and upper limit fixed.
}
\seealso{
\code{\link{gompertz}}, \code{\link{G.3}}, \code{\link{G.4}}
}
\keyword{models}
\keyword{nonlinear}
