% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.R
\name{G.3u}
\alias{G.3u}
\title{Three-parameter Gompertz model with upper limit fixed}
\usage{
G.3u(upper = 1, fixed = c(NA, NA, NA), names = c("b", "c", "e"), ...)
}
\arguments{
\item{upper}{numeric specifying the fixed upper horizontal asymptote. Default is 1.}

\item{fixed}{numeric vector of length 3 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{gompertz}}.}
}
\value{
A list (see \code{\link{gompertz}}).
}
\description{
Convenience function for the Gompertz model with the upper limit fixed.
}
\seealso{
\code{\link{gompertz}}, \code{\link{G.2}}, \code{\link{G.3}}, \code{\link{G.4}}
}
\keyword{models}
\keyword{nonlinear}
