% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.r
\name{gaussian}
\alias{gaussian}
\title{Normal (Gaussian) biphasic dose-response model}
\usage{
gaussian(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText,
  loge = FALSE
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e, f.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}

\item{loge}{logical indicating whether or not e or log(e) should be a parameter in the model.
By default e is a model parameter.}
}
\value{
The value returned is a list containing the nonlinear function, the self starter function
and the parameter names.
}
\description{
Model function for fitting symmetric or skewed bell-shaped/biphasic dose-response patterns
using the Gaussian (normal distribution) model.
}
\seealso{
\code{\link{lgaussian}}, \code{\link{drm}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
