% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.ssf.R
\name{gaussian.ssf}
\alias{gaussian.ssf}
\title{Self-starter for Gaussian model}
\usage{
gaussian.ssf(
  method = c("1", "2", "3", "4"),
  fixed,
  logg = FALSE,
  useFixed = FALSE
)
}
\description{
Self-starter for Gaussian model
}
\keyword{internal}
