% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInitial.R
\name{getInitial}
\alias{getInitial}
\title{Showing starting values used}
\usage{
getInitial(object)
}
\arguments{
\item{object}{object of class 'drc'.}
}
\value{
A vector of starting values for the model parameters used to initialize the
  estimation procedure.
}
\description{
Returns the starting values of the model parameters used when fitting a dose-response model.
}
\note{
This function is masking the standard function in the stats package.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
