% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.ssf.R
\name{gompertz.ssf}
\alias{gompertz.ssf}
\title{Self-starter for Gompertz model}
\usage{
gompertz.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for Gompertz model
}
\keyword{internal}
