% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hewlett.r
\name{hewlett}
\alias{hewlett}
\title{Hewlett Mixture Model}
\usage{
hewlett(
  fixed = c(NA, NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f", "g"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  eps = 1e-10
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value
they are fixed. NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters
(should not contain ":").}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used (optional).}

\item{eps}{numeric tolerance for handling zero dose values.}
}
\value{
A list containing the nonlinear model function, the self starter function,
and the parameter names.
}
\description{
Provides the Hewlett model for describing the joint action of two compounds
in binary mixture experiments. Used internally by \code{\link{mixture}}.
}
\seealso{
\code{\link{mixture}}, \code{\link{voelund}}
}
\author{
Christian Ritz
}
\keyword{internal}
