% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{L.3}
\alias{L.3}
\title{Three-parameter logistic model}
\usage{
L.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3. Specifies which parameters are fixed
and at what value they are fixed. \code{NA} indicates that the corresponding
parameter is not fixed.}

\item{names}{character vector of length 3 giving the names of the parameters
\code{(b, d, e)}. Default is \code{c("b", "d", "e")}.}

\item{...}{additional arguments passed to \code{\link{logistic}}.}
}
\value{
A list of class \code{"Boltzmann"} containing the nonlinear function,
self starter function, and parameter names.
}
\description{
A three-parameter logistic model with the lower limit fixed at 0, given by
\deqn{f(x) = \frac{d}{1 + \exp(b(x - e))}}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = L.3())
}
\seealso{
\code{\link{logistic}}, \code{\link{L.4}}, \code{\link{L.5}}
}
\keyword{models}
\keyword{nonlinear}
