% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.r
\name{L.4}
\alias{L.4}
\title{Four-parameter logistic model}
\usage{
L.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4. Specifies which parameters are fixed
and at what value they are fixed. \code{NA} indicates that the corresponding
parameter is not fixed.}

\item{names}{character vector of length 4 giving the names of the parameters
\code{(b, c, d, e)}. Default is \code{c("b", "c", "d", "e")}.}

\item{...}{additional arguments passed to \code{\link{logistic}}.}
}
\value{
A list of class \code{"Boltzmann"} containing the nonlinear function,
  self starter function, and parameter names.
}
\description{
A four-parameter logistic model (symmetric, with \code{f = 1}), given by
\deqn{f(x) = c + \frac{d - c}{1 + \exp(b(x - e))}}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = L.4())
}
\seealso{
\code{\link{logistic}}, \code{\link{L.3}}, \code{\link{L.5}}
}
\keyword{models}
\keyword{nonlinear}
