% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.r
\name{L.5}
\alias{L.5}
\title{Five-parameter generalized logistic model}
\usage{
L.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 5. Specifies which parameters are fixed
and at what value they are fixed. \code{NA} indicates that the corresponding
parameter is not fixed.}

\item{names}{character vector of length 5 giving the names of the parameters
\code{(b, c, d, e, f)}. Default is \code{c("b", "c", "d", "e", "f")}.}

\item{...}{additional arguments passed to \code{\link{logistic}}.}
}
\value{
A list of class \code{"Boltzmann"} containing the nonlinear function,
  self starter function, and parameter names.
}
\description{
A five-parameter generalized logistic model (asymmetric when \code{f != 1}),
given by
\deqn{f(x) = c + \frac{d - c}{(1 + \exp(b(x - e)))^f}}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = L.5())
}
\seealso{
\code{\link{logistic}}, \code{\link{L.3}}, \code{\link{L.4}}
}
\keyword{models}
\keyword{nonlinear}
