% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin.test.R
\name{lin.test}
\alias{lin.test}
\title{Lack-of-fit test for the mean structure based on cumulated residuals}
\usage{
lin.test(
  object,
  noksSim = 20,
  seed = 20070325,
  plotit = TRUE,
  log = "",
  bp = 0.01,
  xlab,
  ylab,
  ylim,
  ...
)
}
\arguments{
\item{object}{object of class 'drc'.}

\item{noksSim}{numeric specifying the number of simulations used to obtain the p-value.}

\item{seed}{numeric specifying the seed value for the random number generator.}

\item{plotit}{logical indicating whether or not the observed cumulated residual process
should be plotted. Default is to plot the process.}

\item{log}{character string which should contain \code{"x"} if the x axis is to be
logarithmic, \code{"y"} if the y axis is to be logarithmic and \code{"xy"} or
\code{"yx"} if both axes are to be logarithmic. The empty string \code{""} yields
the original axes.}

\item{bp}{numeric value specifying the break point below which the dose is zero.}

\item{xlab}{character string specifying an optional label for the x axis.}

\item{ylab}{character string specifying an optional label for the y axis.}

\item{ylim}{numeric vector of length two, containing the lower and upper limit for the y axis.}

\item{...}{additional arguments to be passed further to the basic \code{\link{plot}} method.}
}
\value{
A p-value for test of the null hypothesis that the mean structure is appropriate.
  Ritz and Martinussen (2009) provide the details.
}
\description{
The function provides a lack-of-fit test for the mean structure based on cumulated
residuals from the model fit.
}
\details{
The function provides a graphical model checking of the mean structure in a dose-response
model. The graphical display is supplemented by a p-value based on a supremum-type test.

The test is applicable even in cases where data are non-normal or exhibit variance
heterogeneity.
}
\examples{
## Fitting a log-logistic model to the dataset 'etmotc'
etmotc.m1<-drm(rgr1~dose1, data=etmotc[1:15,], fct=LL.4())

## Test based on cumulated residuals
lin.test(etmotc.m1, 1000)

}
\references{
Ritz, C and Martinussen, T. (2009) Lack-of-fit tests for assessing mean
  structures for continuous dose-response data, \emph{Submitted manuscript}
}
\seealso{
Other available lack-of-fit tests are the Neill test (\code{\link{neill.test}})
  and ANOVA-based test (\code{\link{modelFit}}).
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
