% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{LL.2}
\alias{LL.2}
\alias{l2}
\title{Two-parameter log-logistic function}
\usage{
LL.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)

l2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)
}
\arguments{
\item{upper}{numeric value, the fixed upper limit (default 1).}

\item{fixed}{numeric vector of length 2, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 2, specifying the names of the
parameters (default: b, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A two-parameter log-logistic function with lower limit fixed at 0 and
upper limit fixed (default 1), primarily for use with binomial/quantal
dose-response data.
}
\details{
The two-parameter log-logistic function is given by the expression
\deqn{f(x) = \frac{upper}{1+\exp(b(\log(x)-\log(e)))}}
}
\examples{
earthworms.m1 <- drm(number/total~dose, weights=total,
  data = earthworms, fct = LL.2(), type = "binomial")

}
\seealso{
\code{\link{LL.3}}, \code{\link{LL.4}}, \code{\link{LL.5}},
  \code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
