% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{LL.3}
\alias{LL.3}
\alias{l3}
\title{Three-parameter log-logistic function}
\usage{
LL.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)

l3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 3, specifying the names of the
parameters (default: b, d, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A three-parameter log-logistic function with lower limit fixed at 0.
}
\details{
The three-parameter log-logistic function is given by the expression
\deqn{f(x) = \frac{d}{1+\exp(b(\log(x)-\log(e)))}}
}
\examples{
ryegrass.model1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.3())

}
\seealso{
\code{\link{LL.2}}, \code{\link{LL.4}}, \code{\link{LL.5}},
  \code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
