% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{LL.3u}
\alias{LL.3u}
\alias{l3u}
\title{Three-parameter log-logistic function with upper limit fixed}
\usage{
LL.3u(upper = 1, fixed = c(NA, NA, NA), names = c("b", "c", "e"), ...)

l3u(upper = 1, fixed = c(NA, NA, NA), names = c("b", "c", "e"), ...)
}
\arguments{
\item{upper}{numeric value, the fixed upper limit (default 1).}

\item{fixed}{numeric vector of length 3, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 3, specifying the names of the
parameters (default: b, c, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A three-parameter log-logistic function with upper limit fixed (default 1),
primarily for use with binomial/quantal dose-response data.
}
\details{
The three-parameter log-logistic function with upper limit fixed is given by
\deqn{f(x) = c + \frac{upper-c}{1+\exp(b(\log(x)-\log(e)))}}
}
\seealso{
\code{\link{LL.2}}, \code{\link{LL.3}}, \code{\link{LL.4}},
\code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
