% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic2.R
\name{LL2.5}
\alias{LL2.5}
\title{Five-Parameter Generalised Log-Logistic Model with log(ED50) as Parameter}
\usage{
LL2.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 5. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that should be
estimated.}

\item{names}{character vector of length 5 giving the names of the
parameters \code{b}, \code{c}, \code{d}, \code{e}, and \code{f}.}

\item{\dots}{additional arguments passed to \code{\link{llogistic2}}.}
}
\value{
A list of class \code{"llogistic"} with the nonlinear function,
self-starter, and related components.
}
\description{
A five-parameter generalised log-logistic model where the ED50 is
parameterised on the log scale. All five parameters (\code{b}, \code{c},
\code{d}, \code{e}, \code{f}) are estimated.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL2.5())

}
\seealso{
\code{\link{llogistic2}}, \code{\link{LL2.3}}, \code{\link{LL2.4}}
}
\keyword{models}
\keyword{nonlinear}
