% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.ssf.R
\name{llogistic.ssf}
\alias{llogistic.ssf}
\title{Self-starter for log-logistic model}
\usage{
llogistic.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for log-logistic model
}
\keyword{internal}
