% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.R
\name{LN.2}
\alias{LN.2}
\title{Two-parameter log-normal dose-response model}
\usage{
LN.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)
}
\arguments{
\item{upper}{numeric specifying the fixed upper horizontal asymptote. Default is 1.}

\item{fixed}{numeric vector of length 2 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{lnormal}}.}
}
\value{
A list (see \code{\link{lnormal}}).
}
\description{
\code{LN.2} is a convenience function for the log-normal model with lower limit fixed at 0
and upper limit fixed (default 1), corresponding to the classic probit model.
}
\seealso{
\code{\link{lnormal}}, \code{\link{LN.3}}, \code{\link{LN.4}}
}
\keyword{models}
\keyword{nonlinear}
