% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.r
\name{LN.3}
\alias{LN.3}
\title{Three-parameter log-normal dose-response model}
\usage{
LN.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{lnormal}}.}
}
\value{
A list (see \code{\link{lnormal}}).
}
\description{
\code{LN.3} is a convenience function for the log-normal model with the lower limit fixed at 0.
}
\seealso{
\code{\link{lnormal}}, \code{\link{LN.2}}, \code{\link{LN.4}}
}
\keyword{models}
\keyword{nonlinear}
