% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.R
\name{LN.3u}
\alias{LN.3u}
\title{Three-parameter log-normal model with upper limit fixed}
\usage{
LN.3u(upper = 1, fixed = c(NA, NA, NA), names = c("b", "c", "e"), ...)
}
\arguments{
\item{upper}{numeric specifying the fixed upper horizontal asymptote. Default is 1.}

\item{fixed}{numeric vector of length 3 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{lnormal}}.}
}
\value{
A list (see \code{\link{lnormal}}).
}
\description{
\code{LN.3u} is a convenience function for the log-normal model with the upper limit fixed (default 1).
}
\seealso{
\code{\link{lnormal}}, \code{\link{LN.2}}, \code{\link{LN.3}}, \code{\link{LN.4}}
}
\keyword{models}
\keyword{nonlinear}
