% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.R
\name{LN.4}
\alias{LN.4}
\title{Four-parameter log-normal dose-response model}
\usage{
LN.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4 specifying fixed parameters (NAs for free parameters).}

\item{names}{character vector of parameter names.}

\item{...}{additional arguments passed to \code{\link{lnormal}}.}
}
\value{
A list (see \code{\link{lnormal}}).
}
\description{
\code{LN.4} is a convenience function for the full four-parameter log-normal model.
}
\seealso{
\code{\link{lnormal}}, \code{\link{LN.2}}, \code{\link{LN.3}}
}
\keyword{models}
\keyword{nonlinear}
