% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.R
\name{lnormal}
\alias{lnormal}
\title{Log-normal dose-response model}
\usage{
lnormal(
  fixed = c(NA, NA, NA, NA),
  names = c("b", "c", "d", "e"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText,
  loge = FALSE
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}

\item{loge}{logical indicating whether or not ED50 or log(ED50) should be a parameter in the model.
By default ED50 is a model parameter.}
}
\value{
A list containing the non-linear function, the self starter function
and the parameter names.
}
\description{
\code{lnormal} provides a general framework for specifying the mean function of the
decreasing or increasing log-normal dose-response model.
}
\details{
For the case where log(ED50) is a parameter in the model, the mean function is:
\deqn{f(x) = c + (d-c)(\Phi(b(\log(x)-e)))}
and in case ED50 is a parameter:
\deqn{f(x) = c + (d-c)(\Phi(b(\log(x)-\log(e))))}

For \eqn{c=0} and \eqn{d=1}, the model reduces to the classic probit model.
}
\references{
Finney, D. J. (1971) \emph{Probit analysis}, London: Cambridge University Press.

Bruce, R. D. and Versteeg, D. J. (1992) A statistical procedure for modeling continuous toxicity data,
\emph{Environ. Toxicol. Chem.}, \bold{11}, 1485--1494.
}
\seealso{
The log-logistic model \code{\link{llogistic}} is very similar to the log-normal model.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
