% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnormal.ssf.R
\name{lnormal.ssf}
\alias{lnormal.ssf}
\title{Self-starter for log-normal model}
\usage{
lnormal.ssf(method = c("1", "2", "3", "4"), fixed, loge, useFixed = FALSE)
}
\description{
Self-starter for log-normal model
}
\keyword{internal}
