% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.ssf.R
\name{logistic.ssf}
\alias{logistic.ssf}
\title{Self-starter for logistic model}
\usage{
logistic.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for logistic model
}
\keyword{internal}
