% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic2.R
\name{lowFixed}
\alias{lowFixed}
\title{Construct Text for Model with Fixed Lower Limit}
\usage{
lowFixed(modelStr)
}
\arguments{
\item{modelStr}{character string with the base model description.}
}
\value{
A character string describing the model with its fixed lower limit.
}
\description{
Helper function that appends lower limit information to a model description
string.
}
\keyword{internal}
