% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{MM.2}
\alias{MM.2}
\title{Two-parameter Michaelis-Menten function}
\usage{
MM.2(fixed = c(NA, NA), names = c("d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 2, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 2, specifying the names of the
parameters (default: d, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A two-parameter Michaelis-Menten function where b is fixed at -1, c at 0,
and f at 1. Commonly used for enzyme kinetics and weed density studies.
}
\details{
The two-parameter Michaelis-Menten function is
\deqn{f(x) = \frac{d \cdot x}{e + x}}
which is equivalent to \eqn{d/(1+(e/x))}.
}
\examples{
met.mm.m1 <- drm(gain~dose, product, data = methionine, fct = MM.2())

}
\seealso{
\code{\link{MM.3}}, \code{\link{AR.2}}, \code{\link{AR.3}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
