% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{MM.3}
\alias{MM.3}
\title{Three-parameter Michaelis-Menten function}
\usage{
MM.3(fixed = c(NA, NA, NA), names = c("c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 3, specifying the names of the
parameters (default: c, d, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A three-parameter (shifted) Michaelis-Menten function where b is fixed
at -1 and f at 1.
}
\details{
The three-parameter Michaelis-Menten function is
\deqn{f(x) = c + \frac{d-c}{1+(e/x)}}
}
\examples{
met.mm.m1 <- drm(gain~dose, product, data = methionine, fct = MM.3())

}
\seealso{
\code{\link{MM.2}}, \code{\link{AR.2}}, \code{\link{AR.3}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
