% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFunction.R
\name{modelFunction}
\alias{modelFunction}
\title{Create model evaluation function}
\usage{
modelFunction(
  dose,
  parm2mat,
  drcFct,
  cm,
  assayNoOld,
  upperPos,
  retFct,
  doseScaling,
  respScaling,
  isFinite,
  pshifts = NULL
)
}
\description{
Create model evaluation function
}
\keyword{internal}
