% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nec.R
\name{NEC}
\alias{NEC}
\title{No Effect Concentration (NEC) dose-response model}
\usage{
NEC(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), fctName, fctText)
}
\arguments{
\item{fixed}{numeric vector specifying which parameters are fixed and at what value they are fixed.
NAs are used for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters (should not contain ":").
The default is reasonable (see under 'Usage').}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list containing the nonlinear function, the self starter function
and the parameter names.
}
\description{
The NEC model is a dose-response model with a threshold below which the response is assumed
constant and equal to the control response. It has been proposed as an alternative to both the
classical NOEC and the regression-based EC/ED approach.
}
\details{
The NEC model function proposed by Pires et al (2002) is:
\deqn{f(x) = c + (d-c)\exp(-b(x-e)I(x-e))}
where \eqn{I(x-e)} is the indicator function equal to 0 for \eqn{x<=e} and 1 for \eqn{x>e}.
}
\examples{
nec.m1 <- drm(rootl ~ conc, data = ryegrass, fct = NEC.4())
summary(nec.m1)
plot(nec.m1)

}
\references{
Pires, A. M., Branco, J. A., Picado, A., Mendonca, E. (2002)
Models for the estimation of a 'no effect concentration',
\emph{Environmetrics}, \bold{13}, 15--27.
}
\seealso{
\code{\link{NEC.2}}, \code{\link{NEC.3}}, \code{\link{NEC.4}}, \code{\link{drm}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
