% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIcompX.R
\name{plotFACI}
\alias{plotFACI}
\title{Plot combination index as a function of fraction affected}
\usage{
plotFACI(
  effList,
  indAxis = c("ED", "EF"),
  caRef = TRUE,
  showPoints = FALSE,
  add = FALSE,
  ylim,
  ...
)
}
\arguments{
\item{effList}{a list as returned by \code{\link{CIcompX}}.}

\item{indAxis}{character string. Either "ED" for effective doses or "EF" for effects.}

\item{caRef}{logical. If TRUE (default), a reference line for concentration addition is drawn.}

\item{showPoints}{logical. If TRUE, estimated combination indices are plotted as points.}

\item{add}{logical. If TRUE, the plot is added to an existing plot.}

\item{ylim}{numeric vector of length 2 giving the range for the y axis.}

\item{...}{additional graphical arguments.}
}
\value{
Invisibly returns the plot matrix of combination index values.
}
\description{
Visualizes the combination index from \code{\link{CIcompX}} as a function of the fraction affected.
}
\seealso{
\code{\link{CIcompX}}, \code{\link{CIcomp}}
}
\author{
Christian Ritz and Ismael Rodea-Palomares
}
\keyword{models}
\keyword{nonlinear}
