% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{PR}
\alias{PR}
\title{Expected or predicted response}
\usage{
PR(object, xVec, ...)
}
\arguments{
\item{object}{object of class \code{drc} obtained from fitting a dose-response model.}

\item{xVec}{numeric vector of dose values.}

\item{...}{additional arguments passed to \code{\link[drc]{predict.drc}}.}
}
\value{
A numeric vector of predicted values or possibly a matrix of predicted values
and corresponding standard errors.
}
\description{
Returns the expected or predicted response for specified dose values. This is a
convenience function for easy access to predicted values.
}
\examples{
ryegrass.m1 <- drm(ryegrass, fct = LL.4())
PR(ryegrass.m1, c(5, 10))

}
\seealso{
\code{\link[drc]{predict.drc}}
}
\author{
Christian Ritz after a suggestion from Andrew Kniss.
}
\keyword{models}
\keyword{nonlinear}
