% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siInner.R
\name{siInner}
\alias{siInner}
\title{Inner function for selectivity index}
\usage{
siInner(
  indPair,
  pVec,
  compMatch,
  object,
  indexMat,
  parmMat,
  varMat,
  level,
  reference,
  type,
  sifct,
  interval,
  degfree,
  logBase
)
}
\description{
Inner function for selectivity index
}
\keyword{internal}
