% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDR.R
\name{simDR}
\alias{simDR}
\title{Simulating ED values under various scenarios}
\usage{
simDR(
  mpar,
  sigma,
  fct,
  noSim = 1000,
  conc,
  edVec = c(10, 50),
  seedVal = 20070723
)
}
\arguments{
\item{mpar}{numeric vector of model parameters.}

\item{sigma}{numeric specifying the residual standard deviation.}

\item{fct}{list supplying the chosen dose-response mean function (e.g., \code{LL.4()}).}

\item{noSim}{numeric giving the number of simulations. Defaults to \code{1000}.}

\item{conc}{numeric vector of concentration/dose values. Must contain at least 5 values.}

\item{edVec}{numeric vector of ED levels to estimate in each simulation. Defaults to
\code{c(10, 50)}.}

\item{seedVal}{numeric giving the seed used to initialise the random number generator.
Defaults to \code{20070723}.}
}
\value{
Invisibly returns a list with one element:
\describe{
\item{\code{se}}{A 3D array of dimensions
\code{(length(conc) - 4) x 6 x length(edVec)} containing empirical
standard deviations of the estimated ED values. Rows correspond to the
number of concentration levels used (starting from 5). Columns correspond
to the number of replicates per concentration (1 to 6). The third dimension
corresponds to each ED level in \code{edVec}.}
}
The array values are also printed to the console during execution.
}
\description{
Simulating ED values for a given model and given dose values.
}
\details{
The arguments \code{mpar} and \code{sigma} are typically obtained from a
previous model fit. Only dose-response models assuming normally distributed
errors can be used.
}
\examples{
ryegrass.m1 <- drm(ryegrass, fct = LL.4())

simDR(
  mpar     = coef(ryegrass.m1),
  sigma    = sqrt(summary(ryegrass.m1)$resVar),
  fct      = LL.4(),
  noSim    = 2,
  conc     = c(1.88, 3.75, 7.50, 0.94, 15, 0.47, 30, 0.23, 60),
  seedVal  = 20070723
)

}
\author{
Christian Ritz, Hannes Reinwald
}
\keyword{models}
\keyword{nonlinear}
