% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDcomp.R
\name{splitInd}
\alias{splitInd}
\title{Split index vectors into shared and unique components}
\usage{
splitInd(ind1, ind2)
}
\description{
Split index vectors into shared and unique components
}
\keyword{internal}
