<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.drc.R
\name{summary.drc}
\alias{summary.drc}
\title{Summarising non-linear model fits}
\usage{
\method{summary}{drc}(object, od = FALSE, pool = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{od}{logical. If TRUE adjustment for over-dispersion is used.}

\item{pool}{logical. If TRUE curves are pooled. Otherwise they are not. This
argument only works for models with independently fitted curves as
specified in \code{\link{drm}}.}

\item{...}{additional arguments.}
}
\value{
A list of summary statistics that includes parameter estimates and
estimated standard errors.
}
\description{
\code{summary} compiles a comprehensive summary for objects of class 'drc'.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())
summary(ryegrass.m1)

}
\seealso{
\code{\link{drm}}, \code{\link{coef.drc}}, \code{\link{confint.drc}}
}
\author{
Christian Ritz
}
=======
\name{summary.drc}

\alias{summary.drc}

\title{Summarising non-linear model fits}

\description{
  'summary' compiles a comprehensive summary for objects of class 'drc'. 
}

\usage{

  \method{summary}{drc}(object, od = FALSE, pool = TRUE, ...)
}

\arguments{
  \item{object}{an object of class 'drc'.}
  \item{od}{logical. If TRUE adjustment for over-dispersion is used.}  
  \item{pool}{logical. If TRUE curves are pooled. Otherwise they are not. This argument only works for models with
  independently fitted curves as specified in \code{\link{drm}}.} 
  \item{...}{additional arguments.}
}

\value{
  A list of summary statistics that includes parameter estimates and estimated standard errors.
}

\author{Christian Ritz}



%\examples{}

>>>>>>> parent of b48f6bb (Delete man directory)
\keyword{models}
\keyword{nonlinear}
