% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threephase.R
\name{threephase}
\alias{threephase}
\title{Three-Phase Dose-Response Model}
\usage{
threephase(
  fixed = c(NA, NA, NA, NA, NA, NA, NA, NA, NA, NA),
  names = c("b1", "c1", "d1", "e1", "b2", "d2", "e2", "b3", "d3", "e3"),
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector specifying which parameters are fixed and at what value
they are fixed. NAs are used for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters
(should not contain ":"). The default is reasonable.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list containing the nonlinear function, the self starter function,
and the parameter names.
}
\description{
A ten-parameter dose-response model combining three log-logistic components,
extending the two-phase model (\code{\link{twophase}}) for describing even more
complex dose-response patterns.
}
\details{
The model function is the sum of a four-parameter log-logistic model and two
three-parameter log-logistic models:

\deqn{f(x) = \mathrm{LL.4}(x; b1, c1, d1, e1) + \mathrm{LL.3}(x; b2, d2, e2) + \mathrm{LL.3}(x; b3, d3, e3)}
}
\seealso{
\code{\link{twophase}}, \code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
