% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{ucedergreen}
\alias{ucedergreen}
\title{U-shaped Cedergreen-Ritz-Streibig model}
\usage{
ucedergreen(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("loglinear", "anke", "method3", "normolle"),
  ssfct = NULL,
  alpha,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{A numeric vector of length 5 specifying any parameters to be held fixed
during the estimation. The order is \code{c(b, c, d, e, f)}. Use \code{NA} for
parameters that should be estimated. The default is to estimate all parameters.}

\item{names}{A character vector of length 5 providing names for the parameters.
The default is \code{c("b", "c", "d", "e", "f")}.}

\item{method}{A character string specifying the method for the self-starter function
to use for finding initial parameter values. Options are \code{"loglinear"},
\code{"anke"}, \code{"method3"}, and \code{"normolle"}. This is only used if \code{ssfct} is \code{NULL}.}

\item{ssfct}{A custom self-starter function. If \code{NULL} (the default), a
self-starter is automatically generated by calling \code{\link{cedergreen.ssf}}
with the specified \code{method}, \code{fixed}, and \code{alpha} arguments.}

\item{alpha}{A mandatory numeric value specifying the fixed shape parameter \eqn{\alpha}.
The function will stop if this is not provided.}

\item{fctName}{An optional character string to name the function object.}

\item{fctText}{An optional character string providing a descriptive text for the model.}
}
\value{
A list of class \code{"UCRS"}, containing the model function (\code{fct}),
the self-starter function (\code{ssfct}), parameter names (\code{names}), and other
components required for use with modeling functions like \code{\link[drc]{drm}}.
}
\description{
\code{ucedergreen} provides a very general way of specifying the Cedergreen-Ritz-Streibig
modified log-logistic model for describing u-shaped hormesis, under various constraints on the parameters.
}
\details{
The u-shaped model is given by the expression
\deqn{f(x) = c + d - \frac{d-c+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}}
}
\references{
Cedergreen, N. and Ritz, C. and Streibig, J. C. (2005)
Improved empirical models describing hormesis,
\emph{Environmental Toxicology and Chemistry} \bold{24}, 3166--3172.
}
\seealso{
\code{\link{cedergreen}}, \code{\link{UCRS.4a}}, \code{\link{UCRS.5a}}, \code{\link{drm}}
}
\author{
Christian Ritz, Hannes Reinwald
}
\keyword{models}
\keyword{nonlinear}
