% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{UCRS.4a}
\alias{UCRS.4a}
\title{U-shaped CRS model with lower limit 0 (alpha=1)}
\usage{
UCRS.4a(names = c("b", "d", "e", "f"), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{ucedergreen}}.}
}
\value{
A list (see \code{\link{ucedergreen}}).
}
\description{
Four-parameter u-shaped CRS hormesis model with lower limit fixed at 0 and alpha=1.
}
\seealso{
\code{\link{ucedergreen}}, \code{\link{UCRS.5a}}, \code{\link{CRS.4a}}
}
\keyword{models}
\keyword{nonlinear}
