% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{UCRS.5a}
\alias{UCRS.5a}
\title{U-shaped CRS five-parameter model (alpha=1)}
\usage{
UCRS.5a(names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{ucedergreen}}.}
}
\value{
A list (see \code{\link{ucedergreen}}).
}
\description{
Five-parameter u-shaped CRS hormesis model with alpha=1.
}
\seealso{
\code{\link{ucedergreen}}, \code{\link{UCRS.4a}}, \code{\link{CRS.5a}}
}
\keyword{models}
\keyword{nonlinear}
