% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml3a}
\alias{uml3a}
\title{Alias for UCRS.4a}
\usage{
uml3a(names = c("b", "d", "e", "f"), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{ucedergreen}}.}
}
\description{
\code{uml3a} is an alias for \code{\link{UCRS.4a}}.
}
\seealso{
\code{\link{UCRS.4a}}
}
\keyword{models}
\keyword{nonlinear}
