% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml4c}
\alias{uml4c}
\title{Alias for UCRS.5c}
\usage{
uml4c(names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{ucedergreen}}.}
}
\description{
\code{uml4c} is an alias for \code{\link{UCRS.5c}}.
}
\seealso{
\code{\link{UCRS.5c}}
}
\keyword{models}
\keyword{nonlinear}
