% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic2.R
\name{upFixed}
\alias{upFixed}
\title{Construct Text for Model with Fixed Upper Limit}
\usage{
upFixed(modelStr, upper)
}
\arguments{
\item{modelStr}{character string with the base model description.}

\item{upper}{numeric value for the fixed upper limit.}
}
\value{
A character string describing the model with its fixed upper limit.
}
\description{
Helper function that appends upper limit information to a model description
string.
}
\keyword{internal}
