% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1.R
\name{W1.4}
\alias{W1.4}
\alias{w4}
\title{Four-parameter Weibull type 1 model}
\usage{
W1.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)

w4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that are not fixed.}

\item{names}{character vector of length 4 giving the names of the
parameters. The default is \code{c("b", "c", "d", "e")}.}

\item{\dots}{additional arguments passed to \code{\link{weibull1}}, most
notably \code{method} (a character string: \code{"1"} (default),
\code{"2"}, \code{"3"}, or \code{"4"}) which selects the self-starter
method for obtaining starting values. See \code{\link{weibull1}} for
details.}
}
\value{
A list of class \code{Weibull-1} containing the nonlinear function,
self starter function, and parameter names.
}
\description{
A four-parameter Weibull type 1 model.
}
\details{
The model is given by the expression
\deqn{f(x) = c + (d - c) \exp(-\exp(b(\log(x) - \log(e))))}
}
\examples{
terbuthylazin.m1 <- drm(rgr ~ dose, data = terbuthylazin, fct = W1.4())

}
\references{
Seber, G. A. F. and Wild, C. J. (1989)
\emph{Nonlinear Regression}, New York: Wiley & Sons (pp. 338--339).

Ritz, C. (2009)
Towards a unified approach to dose-response modeling in ecotoxicology.
\emph{Environ Toxicol Chem}, \bold{29}, 220--229.
}
\seealso{
\code{\link{weibull1}}, \code{\link{W1.2}}, \code{\link{W1.3}}
}
\keyword{models}
\keyword{nonlinear}
