% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{W2.3}
\alias{W2.3}
\title{Three-parameter Weibull (type 2) model}
\usage{
W2.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3, specifying fixed parameters (use \code{NA} for
parameters that should be estimated).}

\item{names}{character vector of length 3 giving the names of the parameters
(default \code{c("b", "d", "e")}).}

\item{...}{additional arguments passed to \code{\link{weibull2}}, most
notably \code{method} (a character string: \code{"1"} (default),
\code{"2"}, \code{"3"}, or \code{"4"}) which selects the self-starter
method for obtaining starting values. See \code{\link{weibull2}} for
details.}
}
\value{
A list of class \code{"Weibull-2"} as returned by \code{\link{weibull2}}.
}
\description{
A three-parameter Weibull type 2 model with the lower limit fixed at 0.
The model is given by the equation
\deqn{f(x) = d \cdot (1 - \exp(-\exp(b(\log(x) - \log(e)))))}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = W2.3())

}
\seealso{
\code{\link{weibull2}}, \code{\link{W2.2}}, \code{\link{W2.4}}
}
\keyword{models}
\keyword{nonlinear}
