% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1.ssf.R
\name{weibull1.ssf}
\alias{weibull1.ssf}
\title{Self-starter for Weibull type 1 model}
\usage{
weibull1.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for Weibull type 1 model
}
\keyword{internal}
