% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.ssf.R
\name{weibull2.ssf}
\alias{weibull2.ssf}
\title{Self-starter for Weibull type 2 model}
\usage{
weibull2.ssf(method = c("1", "2", "3", "4"), fixed, useFixed = FALSE)
}
\description{
Self-starter for Weibull type 2 model
}
\keyword{internal}
