\name{aconiazide}

\alias{aconiazide}

\docType{data}

\title{Weight change in rats after exposure to a medical drug}

\description{
  For each of 4 dose levels the weight change over 6 monts is reported for 14 rats exposed to an antituberculosis drug, aconiazide.
}

\usage{data(aconiazide)}

\format{
  A data frame with 55 observations of the following 2 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{weightChange}}{a numeric vector giving weight change (g) after 6 months of exposure}
    }
}

\source{
Beland, F. A. and Dooley, K. L. and Hansen, E. B. and Sheldon, W. G. (1995). Six-month toxicity comparison of the antituberculosis drugs aconiazide and isoniazid in fischer 344 rats. Journal of the American College of Toxicology, \bold{14(4)}:328--342.
}

\examples{
\dontrun{
library(drc)

## Displaying the data
head(aconiazide)

## Fitting a four-parameter log-logistic model
aconiazide.m1 <- drm(weightChange ~ dose, data = aconiazide, fct = LL.4())
summary(aconiazide.m1)

## Plotting the fitted curve
plot(aconiazide.m1, xlab = "Dose", ylab = "Weight change (g)")
}
}

\keyword{datasets}
