\name{acute.inh}

\alias{acute.inh}

\docType{data}

\title{Acute inhalation}

\description{acute.inh data.}

\usage{data(acute.inh)}

\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
    \item{\code{num.dead}}{a numeric vector}
  }
}




\keyword{datasets}
