% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.drc.R
\name{anova.drc}
\alias{anova.drc}
\title{ANOVA for dose-response model fits}
\usage{
\method{anova}{drc}(object, ..., details = TRUE, test = NULL)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{...}{additional arguments.}

\item{details}{logical indicating whether or not details on the models
compared should be displayed. Default is TRUE (details are displayed).}

\item{test}{a character string specifying the test statistic to be applied.
Use \code{"od"} to assess overdispersion for binomial data.}
}
\value{
An object of class 'anova'.
}
\description{
\code{anova} produces an analysis of variance table for one or two
non-linear model fits.
}
\details{
Specifying only a single object gives a test for lack-of-fit, comparing the
non-linear regression model to a more general one-way or two-way ANOVA
model.

If two objects are specified a test for reduction from the larger to the
smaller model is given. (This only makes statistical sense if the models are
nested, that is: one model is a submodel of the other model.)
}
\examples{
## Comparing Weibull three- and four-parameter models using an F test
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = W1.4())
ryegrass.m2 <- drm(rootl ~ conc, data = ryegrass, fct = W1.3())
anova(ryegrass.m2, ryegrass.m1)

anova(ryegrass.m2, ryegrass.m1, details = FALSE)  # without details

}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
