% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.r
\name{AR.3}
\alias{AR.3}
\title{Three-parameter shifted asymptotic regression model}
\usage{
AR.3(fixed = c(NA, NA, NA), names = c("c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3, specifying fixed parameters (use \code{NA} for
parameters that should be estimated).}

\item{names}{character vector of length 3 giving the names of the parameters
(default \code{c("c", "d", "e")}).}

\item{...}{additional arguments passed to \code{\link{weibull2}}.}
}
\value{
A list of class \code{"Weibull-2"} as returned by \code{\link{weibull2}}.
}
\description{
A three-parameter asymptotic regression model where \code{b} is fixed at 1.
The model is given by the equation
\deqn{f(x) = c + (d - c)(1 - \exp(-x / e))}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = AR.3())

}
\seealso{
\code{\link{AR.2}}, \code{\link{weibull2}}, \code{\link{EXD.3}}
}
\keyword{models}
\keyword{nonlinear}
