% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arandaordaz.R
\name{arandaordaz}
\alias{arandaordaz}
\title{Asymptotic Regression Model}
\usage{
arandaordaz(fixed = c(NA, NA, NA), names = c("a", "b", "c"), fctName, fctText)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value
they are fixed. NAs for parameters that are not fixed.}

\item{names}{vector of character strings giving the names of the parameters
(should not contain ":").}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list of class \code{drcMean}, containing the mean function, the self starter
function, the parameter names, and other components such as derivatives and a function
for calculating ED values.
}
\description{
The base function for the asymptotic regression model, providing the mean
function and self starter for a three-parameter model.
}
\details{
The asymptotic regression model is a three-parameter model with mean function:

\deqn{f(x) = c + (d-c)(1-\exp(-x/e))}

The parameter \eqn{c} is the lower limit (at \eqn{x=0}), \eqn{d} is the upper limit,
and \eqn{e>0} determines the steepness of the increase.
}
\note{
The function is for use with \code{\link{drm}}.
}
\examples{
## Asymptotic regression on methionine data
met.as.m1 <- drm(gain ~ dose, product, data = methionine, fct = AR.3(),
pmodels = list(~1, ~factor(product), ~factor(product)))
plot(met.as.m1, log = "", ylim = c(1450, 1800))
summary(met.as.m1)

}
\seealso{
\code{\link{AR.2}}, \code{\link{AR.3}}, \code{\link{EXD.2}}, \code{\link{EXD.3}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
