\name{arbovirus}

\alias{arbovirus}

\docType{data}

\title{arbovirus}

\description{Data from a dose-response experiment with an arbovirus involving two treatment groups (FP and SP). For each dose level, the total number of subjects and the numbers of dead and defective subjects were recorded.}

\usage{data(arbovirus)}

\format{
  A data frame with 9 observations on the following 5 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
    \item{\code{dead}}{a numeric vector}
    \item{\code{def}}{a numeric vector}
    \item{\code{trt}}{a categorical vector}
  }
}

\keyword{datasets}
