\name{arbovirus}

\alias{arbovirus}

\docType{data}

\title{arbovirus}

\description{arbovirus data.}

\usage{data(arbovirus)}

\format{
  A data frame with 9 observations on the following 5 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
    \item{\code{dead}}{a numeric vector}
    \item{\code{def}}{a numeric vector}
    \item{\code{trt}}{a categorical vector}
  }
}

\keyword{datasets}
