\name{barley}

\alias{barley}

\docType{data}

\title{Barley}

\description{Barley data.}

\usage{data(barley)}

\format{
  A data frame with 18 observations of the following 2 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector}
    \item{\code{weight}}{a numeric vector}
    }
}


\keyword{datasets}
