% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincousens.R
\name{bcl4}
\alias{bcl4}
\title{Alias for BC.5}
\usage{
bcl4(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\description{
\code{bcl4} is an alias for \code{\link{BC.5}}.
}
\seealso{
\code{\link{BC.5}}
}
\keyword{models}
\keyword{nonlinear}
