% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.r
\name{bread.drc}
\alias{bread.drc}
\title{Bread for the sandwich estimator}
\usage{
bread.drc(x, ...)
}
\arguments{
\item{x}{object of class \code{drc}.}

\item{...}{additional arguments. At the moment none are supported.}
}
\value{
The unscaled hessian matrix.
}
\description{
Computes the "bread" (unscaled hessian) for the sandwich estimator of the
variance-covariance matrix for objects of class 'drc'.
}
\details{
The details are provided by Zeileis (2006).
}
\references{
Zeileis, A. (2006) Object-oriented Computation of Sandwich Estimators,
\emph{J. Statist. Software}, \bold{16}, Issue 9.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
