\name{C.dubia}

\alias{C.dubia}

\docType{data}

\title{Offsprings resulting from a toxicity test}

\description{
Results from a toxicity test with seven different concentrations of waste water. Results were reported as number ofsprings of a water flea, C. dubia.
}

\usage{data(C.dubia)}

\format{
  A data frame with 50 observations of the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector giving waste water in percentage}
    \item{\code{number}}{a numeric vector}
  }
}

\source{
A. J. Bailer and J. T. Oris (1997). Estimating inhibition concentrations for different response scales
using generalized linear models. Environmental Toxicology and Chemistry, \bold{16}:1554--1559.
}


\keyword{datasets}
