\name{CadmiumDaphnia}

\alias{CadmiumDaphnia}

\docType{data}

\title{Cadmium Daphnia Data}

\description{
  Cadmium Daphnia Data.
}

\usage{data(CadmiumDaphnia)}

\format{
  A data frame with 58 observations on the following 6 variables.
  \describe{
    \item{\code{Dose}}{a numeric vector of dose values}
    \item{\code{Time}}{a numeric vector}
    \item{\code{Total}}{a numeric vector}
    \item{\code{Start}}{a numeric vector}
    \item{\code{End}}{a numeric vector}
    \item{\code{Dead}}{a numeric vector}
  }
}

\keyword{datasets}
