<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{cedergreen}
\alias{cedergreen}
\title{Cedergreen-Ritz-Streibig Model}
\usage{
cedergreen(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("loglinear", "anke", "method3", "normolle"),
  ssfct = NULL,
  alpha,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{A numeric vector of length 5 specifying any parameters to be held fixed
during the estimation. The order is \code{c(b, c, d, e, f)}. Use \code{NA} for
parameters that should be estimated. The default is to estimate all parameters.}

\item{names}{A character vector of length 5 providing names for the parameters.
The default is \code{c("b", "c", "d", "e", "f")}.}

\item{method}{A character string specifying the method for the self-starter function
to use for finding initial parameter values. Options are \code{"loglinear"},
\code{"anke"}, \code{"method3"}, and \code{"normolle"}. This is only used if \code{ssfct} is \code{NULL}.}

\item{ssfct}{A custom self-starter function. If \code{NULL} (the default), a
self-starter is automatically generated by calling \code{\link{cedergreen.ssf}}
with the specified \code{method}, \code{fixed}, and \code{alpha} arguments.}

\item{alpha}{A mandatory numeric value specifying the fixed shape parameter \eqn{\alpha}.
The function will stop if this is not provided.}

\item{fctName}{An optional character string to name the function object.}

\item{fctText}{An optional character string providing a descriptive text for the model.}
}
\value{
A list of class \code{mllogistic}, containing the model function (\code{fct}),
the self-starter function (\code{ssfct}), parameter names (\code{names}), and other
components required for use with modeling functions like \code{\link[drc]{drm}}.
}
\description{
Provides the Cedergreen-Ritz-Streibig function, a five-parameter model
for describing dose-response curves that exhibit hormesis (a stimulatory or
beneficial effect at low doses). This function generates a model object suitable
for use with non-linear regression functions like \code{\link[drc]{drm}}.
}
\details{
The Cedergreen-Ritz-Streibig model is defined by the following equation:
\deqn{f(x) = c + \frac{d - c + f \exp(-1/x^{\alpha})}{1 + \exp(b(\log(x) - \log(e)))}}
The parameter \eqn{f} determines the size of the hormetic effect (stimulation).
If \eqn{f=0}, the model simplifies to the standard four-parameter log-logistic model.
The parameter \eqn{\alpha} is a shape parameter that must be specified by the user.
}
\examples{
\donttest{
# This example requires the 'drc' package to be installed.
if (requireNamespace("drc", quietly = TRUE)) {

  library(drc)

  # Sample data exhibiting a hormetic effect (initial stimulation)
  dose <- c(0, 0.1, 0.5, 1, 5, 10, 20)
  response <- c(100, 102, 95, 80, 40, 25, 20)
  my_data <- data.frame(dose = dose, response = response)

  # 1. Basic model fit, specifying the mandatory 'alpha' parameter
  # The self-starter will find initial values for all 5 parameters.
  model_fit <- drm(response ~ dose, data = my_data,
                   fct = cedergreen(alpha = 0.5))
  
  summary(model_fit)
  plot(model_fit, main = "Cedergreen Model Fit")

  # 2. Model fit with a fixed lower limit (c = 20)
  # The self-starter will now only find initial values for b, d, e, and f.
  model_fit_fixed <- drm(response ~ dose, data = my_data,
                         fct = cedergreen(alpha = 0.5, fixed = c(NA, 20, NA, NA, NA)))
        
  summary(model_fit_fixed)
}
}

}
\seealso{
\code{\link[drc]{drm}} for model fitting, and \code{\link{cedergreen.ssf}} for the
underlying self-starter function.
}
=======
\name{cedergreen}

\alias{cedergreen}
\alias{CRS.6}
\alias{ucedergreen}

\title{The Cedergreen-Ritz-Streibig model}

\description{
  'cedergreen' provides a very general way of specifying then Cedergreen-Ritz-Streibig 
  modified log-logistic model for describing hormesis, under various constraints on the parameters.

  \code{\link{CRS.6}} is the extension of \code{link{cedergreen}} with freely varying alpha parameter.

  For u-shaped hormesis data 'ucedergreen' provides a very general way of specifying the 
  Cedergreen-Ritz-Streibig modified log-logistic model, under various constraints on the parameters.
}

\usage{
  cedergreen(fixed = c(NA, NA, NA, NA, NA), 
  names = c("b", "c", "d", "e", "f"), 
  method = c("1", "2", "3", "4"), ssfct = NULL,
  alpha, fctName, fctText)
  
  CRS.6(fixed = c(NA, NA, NA, NA, NA, NA), 
  names = c("b","c","d","e","f","g"),
  method = c("1", "2", "3", "4"), ssfct = NULL)
  
  ucedergreen(fixed = c(NA, NA, NA, NA, NA), 
  names = c("b", "c", "d", "e", "f"), 
  method = c("1", "2", "3", "4"), ssfct = NULL,
  alpha)
}

\arguments{
  \item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed. 
  NAs for parameter that are not fixed.}
  \item{names}{a vector of character strings giving the names of the parameters (should not contain ":"). 
  The order of the parameters is: b, c, d, e, f (see under 'Details').}
  \item{method}{character string indicating the self starter function to use.}  
  \item{ssfct}{a self starter function to be used.}                 
  \item{alpha}{numeric value between 0 and 1, reflecting the steepness of the hormesis peak. 
  This argument needs to be specified.}
  \item{fctName}{optional character string used internally by convenience functions.}  
  \item{fctText}{optional character string used internally by convenience functions.}  
}

\details{
The model is given by the expression
  
  \deqn{ f(x) = c + \frac{d-c+f exp(-1/(x^{\alpha}))}{1+exp(b(log(x)-log(e)))}}
  
which is a five-parameter model (alpha is fixed or freely varying). Not all features (eg EC/ED calculation) 
are available for the model with freely varying alpha.
  
It is a modification of the four-parameter logistic curve to take hormesis into account.
  
The u-shaped model is given by the expression
  
  \deqn{ f(x) = cd - \frac{d-c+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}}
  
}

\value{
  The value returned is a list containing the non-linear function, the self starter function
  and the parameter names.
}

\references{
  Cedergreen, N. and Ritz, C. and Streibig, J. C. (2005) 
  Improved empirical models describing hormesis, 
  \emph{Environmental Toxicology and Chemistry} 
  \bold{24}, 3166--3172.
}

\author{Christian Ritz}

\note{
  The functions are for use with the functions \code{\link{drm}}.
}

\seealso{
  For fixed alpha, several special cases are handled by the following convenience functions 
  \code{\link{CRS.4a}}, \code{\link{CRS.4b}}, 
  \code{\link{CRS.4c}}, \code{\link{CRS.5a}}, \code{\link{CRS.5b}}, \code{\link{CRS.5c}},
  \code{\link{UCRS.4a}}, \code{\link{UCRS.4b}}, \code{\link{UCRS.4c}}, \code{\link{UCRS.5a}},
  \code{\link{UCRS.5b}}, \code{\link{UCRS.5c}} where a, b and c correspond to 
  the pre-specified alpha values 1, 0.5 and 0.25, respectively.
}

\examples{

## Estimating CRS model with alpha unknown
lettuce.crsm1 <- drm(weight~conc, data = lettuce, fct = CRS.6())
summary(lettuce.crsm1)
plot(lettuce.crsm1)  # oops: not increasing until hormesis peak

}

\keyword{models}
\keyword{nonlinear}

\concept{hormesis hormetic effect initial stimulation u-shaped}
>>>>>>> parent of b48f6bb (Delete man directory)
