\name{chlorac}

\alias{chlorac}

\docType{data}

\title{chlorac}

\description{chlorac data.}

\usage{data(chlorac)}

\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{conc}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
    \item{\code{num.dead}}{a numeric vector}
  }
}


\keyword{datasets}
