\name{chlordan}

\alias{chlordan}

\docType{data}

\title{Chlordan}

\description{Data from a chronic toxicity test measuring the reproduction of \emph{Daphnia} exposed to different concentrations of chlordane at two time points. The response measured was the number of offspring (repro) per replicate.}

\usage{data(chlordan)}

\format{
  A data frame with 60 observations on the following 5 variables.
  \describe{
    \item{\code{replicate}}{a numeric vector}
    \item{\code{conc}}{a numeric vector}
    \item{\code{repro}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
  }
}

\keyword{datasets}
