% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.drc.R
\name{coef.drc}
\alias{coef.drc}
\title{Extract Model Coefficients}
\usage{
\method{coef}{drc}(object, ...)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{...}{additional arguments.}
}
\value{
A vector of parameter coefficients which are extracted from the
model object \code{object}.
}
\description{
Extract parameter estimates.
}
\examples{
## Fitting a four-parameter log-logistic model
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())
coef(ryegrass.m1)

}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
