% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{CRS.4a}
\alias{CRS.4a}
\title{Cedergreen-Ritz-Streibig model with lower limit 0 (alpha=1)}
\usage{
CRS.4a(names = c("b", "c", "d", "e", "f"), fixed = c(NA, 0, NA, NA, NA), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{fixed}{a numeric vector specifying which parameters are fixed and at
what value. \code{NA} indicates a parameter to be estimated. The default
fixes the lower limit (\code{c}) at 0 and estimates all other parameters.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\value{
A list (see \code{\link{cedergreen}}).

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated as of version 3.3.0. Please use \code{\link[=CRS.5]{CRS.5()}} instead.
}
\description{
Four-parameter CRS hormesis model with the lower limit fixed at 0 and alpha=1.
}
\examples{
lettuce.crsm1 <- drm(lettuce[,c(2,1)], fct = CRS.4a())
summary(lettuce.crsm1)
ED(lettuce.crsm1, c(50))

}
\seealso{
\code{\link{cedergreen}}, \code{\link{CRS.5a}}, \code{\link{UCRS.4a}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
