% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{CRS.4b}
\alias{CRS.4b}
\title{Cedergreen-Ritz-Streibig model with lower limit 0 (alpha=0.5)}
\usage{
CRS.4b(names = c("b", "c", "d", "e", "f"), fixed = c(NA, 0, NA, NA, NA), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\value{
A list (see \code{\link{cedergreen}}).

`r lifecycle::badge("deprecated")` 
This function is deprecated as of version 3.3.0. Please use [CRS.5()] instead.
}
\description{
Four-parameter CRS hormesis model with the lower limit fixed at 0 and alpha=0.5.
}
\seealso{
\code{\link{cedergreen}}, \code{\link{CRS.4a}}, \code{\link{CRS.5b}}
}
\keyword{models}
\keyword{nonlinear}
