% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{CRS.5a}
\alias{CRS.5a}
\title{Cedergreen-Ritz-Streibig five-parameter model (alpha=1)}
\usage{
CRS.5a(names = c("b", "c", "d", "e", "f"), fixed = c(NA, NA, NA, NA, NA), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\value{
A list (see \code{\link{cedergreen}}).

`r lifecycle::badge("deprecated")` 
This function is deprecated as of version 3.3.0. Please use [CRS.5()] instead.
}
\description{
Five-parameter CRS hormesis model with alpha=1.
}
\examples{
lettuce.m1 <- drm(lettuce[,c(2,1)], fct = CRS.5a())
summary(lettuce.m1)
ED(lettuce.m1, c(50))

}
\seealso{
\code{\link{cedergreen}}, \code{\link{CRS.4a}}, \code{\link{UCRS.5a}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
