% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{CRS.5b}
\alias{CRS.5b}
\title{Cedergreen-Ritz-Streibig five-parameter model (alpha=0.5)}
\usage{
CRS.5b(names = c("b", "c", "d", "e", "f"), fixed = c(NA, NA, NA, NA, NA), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\value{
A list (see \code{\link{cedergreen}}).

`r lifecycle::badge("deprecated")` 
This function is deprecated as of version 3.3.0. Please use [CRS.5()] instead.
}
\description{
Five-parameter CRS hormesis model with alpha=0.5.
}
\seealso{
\code{\link{cedergreen}}, \code{\link{CRS.4b}}, \code{\link{CRS.5a}}
}
\keyword{models}
\keyword{nonlinear}
