% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{CRS.5c}
\alias{CRS.5c}
\title{Cedergreen-Ritz-Streibig five-parameter model (alpha=0.25)}
\usage{
CRS.5c(names = c("b", "c", "d", "e", "f"), fixed = c(NA, NA, NA, NA, NA), ...)
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\value{
A list (see \code{\link{cedergreen}}).

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated as of version 3.3.0. Please use \code{\link[=CRS.5]{CRS.5()}} instead.
}
\description{
Five-parameter CRS hormesis model with alpha=0.25.
}
\seealso{
\code{\link{cedergreen}}, \code{\link{CRS.4c}}, \code{\link{CRS.5a}}
}
\keyword{models}
\keyword{nonlinear}
