% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc-package.R
\docType{package}
\name{drc-package}
\alias{drc}
\alias{drc-package}
\title{drc: Analysis of Dose-Response Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analysis of dose-response data is made available through a suite of flexible and versatile model fitting and after-fitting functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hreinwald.github.io/drc}
  \item \url{https://github.com/hreinwald/drc}
  \item \url{https://www.bioassay.dk}
  \item \url{https://www.r-project.org}
  \item \url{https://cran.r-project.org/web/packages/drc/index.html}
  \item Report bugs at \url{https://github.com/hreinwald/drc/issues/}
}

}
\author{
\strong{Maintainer}: Hannes Reinwald \email{hannes.reinwald@bayer.com}

Authors:
\itemize{
  \item Christian Ritz \email{ritz@bioassay.dk}
  \item Jens C. Streibig \email{streibig@bioassay.dk}
}

}
\keyword{internal}
