% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED.drc.R
\name{ED.drc}
\alias{ED.drc}
\title{Estimating effective doses}
\usage{
\method{ED}{drc}(
  object,
  respLev = c(10, 20, 50),
  interval = c("none", "delta", "fls", "tfls", "inv"),
  clevel = NULL,
  level = 0.95,
  reference = c("control", "upper"),
  type = c("relative", "absolute"),
  lref,
  uref,
  bound = TRUE,
  vcov. = vcov,
  display = TRUE,
  logBase = NULL,
  multcomp = FALSE,
  intType = "confidence",
  ...
)
}
\arguments{
\item{object}{an object of class \code{drc}.}

\item{respLev}{a numeric vector containing the response levels.}

\item{interval}{character string specifying the type of confidence intervals
to be supplied. The default is \code{"none"}. See Details below for more
explanation.}

\item{clevel}{character string specifying the curve id in case estimates for
a specific curve or compound are requested. By default estimates are shown
for all curves.}

\item{level}{numeric. The level for the confidence intervals. Must be a
single value strictly between 0 and 1. The default is \code{0.95}.}

\item{reference}{character string. Is the upper limit or the control level
the reference?}

\item{type}{character string. Whether the specified response levels are
absolute or relative (default).}

\item{lref}{numeric value specifying the lower limit to serve as reference.}

\item{uref}{numeric value specifying the upper limit to serve as reference
(e.g., 100\\%).}

\item{bound}{logical. Default is \code{TRUE}, in which case only ED values
between 0 and 100\\% are allowed. Set to \code{FALSE} for hormesis models.}

\item{vcov.}{function providing the variance-covariance matrix, or a
variance-covariance matrix directly. \code{\link{vcov}} is the default,
but \code{sandwich} is also an option for obtaining robust standard errors.}

\item{display}{logical. If \code{TRUE} results are displayed. Otherwise they
are not (useful in simulations).}

\item{logBase}{numeric. The base of the logarithm in case logarithm
transformed dose values are used.}

\item{multcomp}{logical to switch on output for use with the package
\pkg{multcomp} (which needs to be activated first). Default is
\code{FALSE}.}

\item{intType}{string specifying the type of interval to use with the
predict method in case the type of confidence interval chosen is inverse
regression.}

\item{...}{additional arguments passed to the ED function in the model.}
}
\value{
An invisible matrix containing the estimates and the corresponding
estimated standard errors and possibly lower and upper confidence limits.
Or, alternatively, a list with elements that may be plugged directly into
\code{parm} in the package \pkg{multcomp} (when \code{multcomp = TRUE}).
}
\description{
\code{ED} estimates effective concentration or doses for one or more
specified absolute or relative response levels.
}
\details{
There are several options for calculating confidence intervals through the
argument \code{interval}. The option \code{"delta"} results in asymptotical
Wald-type confidence intervals (using the delta method and the normal or
t-distribution depending on the type of response). The option \code{"fls"}
produces (possibly skewed) confidence intervals through back-transformation
from the logarithm scale (only meaningful in case the parameter in the model
is log(ED50) as for the \code{\link{llogistic2}} models). The option
\code{"tfls"} is for transforming back and forth from log scale
(experimental). The option \code{"inv"} results in confidence intervals
obtained through inverse regression.

For hormesis models (\code{\link{braincousens}} and
\code{\link{cedergreen}}), the additional arguments \code{lower} and
\code{upper} may be supplied. These arguments specify the lower and upper
limits of the bisection method used to find the ED values.
}
\examples{
## Fitting a 4-parameter log-logistic model
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.4())

## Calculating EC/ED values
ED(ryegrass.m1, c(10, 50, 90))

## Displaying 95\% confidence intervals using the delta method
ED(ryegrass.m1, c(10, 50, 90), interval = "delta")

## Displaying 95\% confidence intervals using back-transformation
ED(ryegrass.m1, c(10, 50, 90), interval = "fls")

## Displaying 95\% confidence intervals using inverse regression
ED(ryegrass.m1, c(10, 50, 90), interval = "inv")

}
\seealso{
\code{\link{EDcomp}} for estimating differences and ratios of ED
values, \code{\link{compParm}} for comparing other model parameters, and
\code{\link{backfit}}.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
