% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED.drc.R
\name{ED}
\alias{ED}
\title{Estimating effective doses}
\usage{
ED(object, ...)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{...}{additional arguments passed to methods.}
}
\description{
Estimates effective concentration or doses for specified response levels.
}
\keyword{models}
\keyword{nonlinear}
