% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDcomp.R
\name{EDcomp}
\alias{EDcomp}
\title{Comparison of relative potencies between dose-response curves}
\usage{
EDcomp(
  object,
  percVec,
  percMat = NULL,
  compMatch = NULL,
  od = FALSE,
  vcov. = vcov,
  reverse = FALSE,
  interval = c("none", "delta", "fieller", "fls"),
  level = ifelse(!(interval == "none"), 0.95, NULL),
  reference = c("control", "upper"),
  type = c("relative", "absolute"),
  display = TRUE,
  pool = TRUE,
  logBase = NULL,
  multcomp = FALSE,
  ...
)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{percVec}{a numeric vector of dosage values.}

\item{percMat}{a matrix with 2 columns providing the pairs of indices of \code{percVec} to be
compared. By default all pairs are compared.}

\item{compMatch}{an optional character vector of names of assays to be compared. If not specified
all comparisons are supplied.}

\item{od}{logical. If TRUE adjustment for over-dispersion is used. This argument only makes a
difference for binomial data.}

\item{vcov.}{function providing the variance-covariance matrix. \code{\link{vcov}} is the default,
but \code{sandwich} is also an option (for obtaining robust standard errors).}

\item{reverse}{logical. If TRUE the order of comparison of two curves is reversed.}

\item{interval}{character string specifying the type of confidence intervals to be supplied.
The default is \code{"none"}. Use \code{"delta"} for asymptotics-based confidence intervals,
\code{"fieller"} for confidence intervals based on Fieller's theorem, or \code{"fls"} for
confidence intervals back-transformed from logarithm scale.}

\item{level}{numeric. The level for the confidence intervals. Default is 0.95.}

\item{reference}{character string. Is the upper limit or the control level the reference?}

\item{type}{character string specifying whether absolute or relative response levels are supplied.}

\item{display}{logical. If TRUE results are displayed. Otherwise they are not (useful in simulations).}

\item{pool}{logical. If TRUE curves are pooled. Otherwise they are not. This argument only works
for models with independently fitted curves as specified in \code{\link{drm}}.}

\item{logBase}{numeric. The base of the logarithm in case logarithm transformed dose values are used.}

\item{multcomp}{logical to switch on output for use with the package \pkg{multcomp}. Default is FALSE.}

\item{...}{additional arguments passed to the function doing the calculations.}
}
\value{
An invisible matrix containing the estimates and the corresponding estimated standard
  errors and possibly lower and upper confidence limits. Or, alternatively, a list with elements
  that may be plugged directly into \code{parm} in the package \pkg{multcomp} (when \code{multcomp}
  is TRUE).
}
\description{
Relative potencies (also called selectivity indices) for arbitrary doses are compared between
fitted dose-response curves.
}
\details{
Fieller's theorem is incorporated using the formulas provided by Kotz and Johnson (1983) and
Finney (1978).

For objects of class 'braincousens' or 'mlogistic' the additional argument may be the 'upper'
argument or the 'interval' argument specifying limits for the bisection method.
}
\examples{
spinach.LL.4 <- drm(SLOPE~DOSE, CURVE, data = spinach, fct = LL.4())

EDcomp(spinach.LL.4, c(50, 50))
EDcomp(spinach.LL.4, c(10, 50))
EDcomp(spinach.LL.4, c(10, 50), reverse = TRUE)

}
\seealso{
\code{\link{ED.drc}} for calculating effective doses.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
