% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1.r
\name{EXD.3}
\alias{EXD.3}
\title{Three-parameter exponential decay model}
\usage{
EXD.3(fixed = c(NA, NA, NA), names = c("c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that are not fixed.}

\item{names}{character vector of length 3 giving the names of the
parameters. The default is \code{c("c", "d", "e")}.}

\item{\dots}{additional arguments passed to \code{\link{weibull1}}.}
}
\value{
A list of class \code{Weibull-1} containing the nonlinear function,
  self starter function, and parameter names.
}
\description{
A three-parameter exponential decay model with the slope parameter \code{b}
fixed at 1.
}
\details{
The model is given by the expression
\deqn{f(x) = c + (d - c) \exp(-x/e)}

This is a special case of the Weibull type 1 model
(\code{\link{weibull1}}) with the slope fixed at 1.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = EXD.3())

}
\references{
Seber, G. A. F. and Wild, C. J. (1989)
\emph{Nonlinear Regression}, New York: Wiley \& Sons (pp. 338--339).
}
\seealso{
\code{\link{EXD.2}}, \code{\link{AR.2}}, \code{\link{AR.3}},
  \code{\link{weibull1}}
}
\keyword{models}
\keyword{nonlinear}
