\name{GiantKelp}

\alias{GiantKelp}

\docType{data}

\title{Measurements of germination tubes for Giant Kelp}

\description{
   Giant kelp, Mactocystis pyrifera, was exposed to 8 different concentrations of copper and the respponse measured was the length of the germination tube.
}

\usage{data(GiantKelp)}

\format{
  A data frame with 39 observations of the following 2 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{tubeLength}}{a numeric vector giving the length of the germination tube (mm)}
    }
}

\source{
G. A. Chapman, D. L. Denton, and J. M. Lazorchak (1995). Short-term methods for estimating
the chronic toxicity of effluents and receiving waters to west coast marine and estuarine
organisms. 
}


\keyword{datasets}
