% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz.r
\name{gompertz}
\alias{gompertz}
\title{Gompertz dose-response or growth curve model}
\usage{
gompertz(
  fixed = c(NA, NA, NA, NA),
  names = c("b", "c", "d", "e"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally by convenience functions.}

\item{fctText}{optional character string used internally by convenience functions.}
}
\value{
A list containing the non-linear function, the self starter function
  and the parameter names.
}
\description{
Provides a very general way of specifying the mean function of the decreasing or increasing
Gompertz dose-response or growth curve models.
}
\details{
The Gompertz model is given by the mean function
\deqn{f(x) = c + (d-c)(\exp(-\exp(b(x-e))))}

If \eqn{b<0} the mean function is increasing; it is decreasing for \eqn{b>0}.
}
\references{
Seber, G. A. F. and Wild, C. J. (1989) \emph{Nonlinear Regression}, New York: Wiley & Sons (p. 331).
}
\seealso{
The Weibull model \code{\link{weibull2}} is closely related to the Gompertz model.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
