\name{guthion}

\alias{guthion}

\docType{data}

\title{guthion}

\description{guthion data.}

\usage{data(guthion)}

\format{
  A data frame with 6 observations on the following 6 variables.
  \describe{
    \item{\code{trt}}{a categorial vector}
    \item{\code{dose}}{a numeric vector}
    \item{\code{alive}}{a numeric vector}
    \item{\code{moribund}}{a numeric vector}
    \item{\code{dead}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
  }
}

\keyword{datasets}
