\name{lemna}

\alias{lemna}

\docType{data}

\title{Lemna}

\description{Data from a dose-response experiment with the aquatic plant \emph{Lemna minor} (duckweed). The response measured was the frond number (count) at different concentrations of a test substance.}

\usage{data(lemna)}

\format{
  A data frame with 44 observations on the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector containing the concentration.}
    \item{\code{frond.num}}{a numeric vector containing the response (count).}
    }
}

\keyword{datasets}
