\name{lemna}

\alias{lemna}

\docType{data}

\title{Lemna}

\description{ Data are from .}

\usage{data(lemna)}

\format{
  A data frame with 44 observations on the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector containing the concentration.}
    \item{\code{frond.num}}{a numeric vector containing the response (count).}
    }
}

\keyword{datasets}
