\name{liver.tumor}

\alias{liver.tumor}

\docType{data}

\title{Liver tumor incidence}

\description{
  Liver tumor incidence in female Sprague-Dawley rats exposed to the chemical like 2,3,7,8-tetrachlorodibenzo-pdioxin
(TCDD).
}

\usage{data(liver.tumor)}

\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{conc}}{a numeric vector reporting the concentration of TCDD (ng/kg)}
    \item{\code{total}}{a numeric vector}
    \item{\code{incidence}}{a numeric vector}
  }
}

%\details{}

\source{
   National Toxicology Program. NTP technical report on the toxicology and carcinogenesis
studies of 2,3,7,8-tetrachlorodibenzo-p-dioxin (tcdd) (CAS No. 1746-01-6) in female Harlan
Sprague-Dawley rats (gavage studies). National Toxicology Program technical report series,
(521):4--232, apr 2006.
}

%\references{}

\examples{
\dontrun{
library(drc)

## Displaying the data
head(liver.tumor)

## Fitting a two-parameter log-logistic model for binomial response
liver.tumor.m1 <- drm(incidence/total ~ conc, weights = total,
data = liver.tumor, fct = LL.2(), type = "binomial")
summary(liver.tumor.m1)

## Plotting the fitted curve
plot(liver.tumor.m1, xlab = "Concentration of TCDD (ng/kg)",
ylab = "Tumor incidence")
}
}


\keyword{datasets}
