% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{LL.4}
\alias{LL.4}
\alias{l4}
\title{Four-parameter log-logistic function}
\usage{
LL.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)

l4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 4, specifying the names of the
parameters (default: b, c, d, e).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A four-parameter log-logistic function.
}
\details{
The four-parameter log-logistic function is given by the expression
\deqn{f(x) = c + \frac{d-c}{1+\exp(b(\log(x)-\log(e)))}}
}
\examples{
spinach.m1 <- drm(SLOPE~DOSE, CURVE, data = spinach, fct = LL.4())

}
\seealso{
\code{\link{LL.3}}, \code{\link{LL.5}}, \code{\link{llogistic}}
}
\author{
Christian Ritz and Jens C. Streibig
}
\keyword{models}
\keyword{nonlinear}
