% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{LL.5}
\alias{LL.5}
\alias{l5}
\title{Five-parameter log-logistic function}
\usage{
LL.5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)

l5(fixed = c(NA, NA, NA, NA, NA), names = c("b", "c", "d", "e", "f"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 5, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 5, specifying the names of the
parameters (default: b, c, d, e, f).}

\item{...}{additional arguments to \code{\link{llogistic}}.}
}
\value{
See \code{\link{llogistic}}.
}
\description{
A five-parameter (generalized) log-logistic function. The function is
asymmetric when f differs from 1.
}
\details{
The five-parameter log-logistic function is given by the expression
\deqn{f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-\log(e))))^f}}
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = LL.5())

}
\seealso{
\code{\link{LL.3}}, \code{\link{LL.4}}, \code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
