% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llogistic.R
\name{llogistic}
\alias{llogistic}
\title{The log-logistic function}
\usage{
llogistic(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector of length 5, specifying fixed parameters
(use NA for non-fixed parameters).}

\item{names}{character vector of length 5, specifying the names of the
parameters: b, c, d, e, f.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{fctName}{optional character string used internally.}

\item{fctText}{optional character string used internally.}
}
\value{
A list containing the nonlinear function, the self starter function,
  and the parameter names.
}
\description{
A very general way of specifying log-logistic models under various
constraints on parameters.
}
\details{
The five-parameter log-logistic function is given by the expression
\deqn{f(x) = c + \frac{d-c}{(1+\exp(b(\log(x)-\log(e))))^f}}
}
\references{
Finney, D. J. (1979).

  Seber, G. A. F. and Wild, C. J. (1989).
}
\seealso{
\code{\link{LL.2}}, \code{\link{LL.3}}, \code{\link{LL.4}},
  \code{\link{LL.5}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
