% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{The general asymmetric five-parameter logistic model}
\usage{
logistic(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector of length 5. Specifies which parameters are fixed
and at what value they are fixed. \code{NA} indicates that the corresponding
parameter is not fixed.}

\item{names}{character vector of length 5 giving the names of the parameters
\code{(b, c, d, e, f)}. Default is \code{c("b", "c", "d", "e", "f")}.}

\item{method}{character string indicating the self starter function to use
(\code{"1"}, \code{"2"}, \code{"3"}, or \code{"4"}).}

\item{ssfct}{a self starter function to be used. If \code{NULL} (default),
a built-in self starter is selected via \code{method}.}

\item{fctName}{optional character string used internally to overwrite the
function name.}

\item{fctText}{optional character string used internally to overwrite the
description text.}
}
\value{
A list of class \code{"Boltzmann"} containing the nonlinear function,
self starter function, and parameter names.
}
\description{
The five-parameter logistic model given by the expression
\deqn{f(x) = c + \frac{d - c}{(1 + \exp(b(x - e)))^f}}
}
\details{
This model differs from the log-logistic in that it uses \code{x} directly
rather than \code{log(x)}. It is sometimes referred to as the Boltzmann model.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = L.4())
}
\seealso{
\code{\link{L.3}}, \code{\link{L.4}}, \code{\link{L.5}},
\code{\link{llogistic}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
