% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{MAX}
\alias{MAX}
\title{Maximum mean response}
\usage{
MAX(object, lower = 0.001, upper = 1000, pool = TRUE)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{lower}{numeric. Lower limit for bisection method. Must be smaller than the EDx level to be calculated.}

\item{upper}{numeric. Upper limit for bisection method. Must be larger than the EDx level to be calculated.}

\item{pool}{logical. If TRUE curves are pooled. Otherwise they are not. This argument only works for
models with independently fitted curves as specified in \code{\link{drm}}.}
}
\value{
A matrix with one row per curve in the data set and two columns: one containing the dose
at which the maximum occurs and one containing the corresponding maximum response.
}
\description{
Estimates the maximum mean response and the dose at which it occurs. This function is only
implemented for the built-in functions of class \code{\link{braincousens}} and \code{\link{cedergreen}}.
}
\examples{
## Fitting a Cedergreen-Ritz-Streibig model
lettuce.m1 <- drm(weight~conc, data = lettuce, fct = CRS.4c())

## Finding maximum average response and the corresponding dose
MAX(lettuce.m1)

}
\references{
Cedergreen, N. and Ritz, C. and Streibig, J. C. (2005) Improved empirical models
describing hormesis, \emph{Environmental Toxicology and Chemistry} \bold{24}, 3166--3172.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
