% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{ml3a}
\alias{ml3a}
\title{Alias for CRS.4a}
\usage{
ml3a(names = c("b", "c", "d", "e", "f"), fixed = c(NA, 0, NA, NA, NA), ...)
}
\description{
\code{ml3a} is an alias for \code{\link{CRS.4a}}.
}
\seealso{
\code{\link{CRS.4a}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated as of version 3.3.0. Please use \code{\link[=CRS.5]{CRS.5()}} instead.
}
\keyword{models}
\keyword{nonlinear}
