% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{ml3b}
\alias{ml3b}
\title{Alias for CRS.4b}
\usage{
ml3b(names = c("b", "c", "d", "e", "f"), fixed = c(NA, 0, NA, NA, NA), ...)
}
\description{
\code{ml3b} is an alias for \code{\link{CRS.4b}}.
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{fixed}{a numeric vector specifying which parameters are fixed and at
what value. \code{NA} indicates a parameter to be estimated. The default
fixes the lower limit (\code{c}) at 0.}

\item{...}{additional arguments passed to \code{\link{cedergreen}}.}
}
\seealso{
\code{\link{CRS.4b}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated as of version 3.3.0. Please use \code{\link[=CRS.5]{CRS.5()}} instead.
}
\keyword{models}
\keyword{nonlinear}
