% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{ml3b}
\alias{ml3b}
\title{Alias for CRS.4b}
\usage{
ml3b(names = c("b", "c", "d", "e", "f"), fixed = c(NA, 0, NA, NA, NA), ...)
}
\description{
\code{ml3b} is an alias for \code{\link{CRS.4b}}.
}
\seealso{
\code{\link{CRS.4b}}
`r lifecycle::badge("deprecated")` 
This function is deprecated as of version 3.3.0. Please use [CRS.5()] instead.
}
\keyword{models}
\keyword{nonlinear}
