% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{ml4b}
\alias{ml4b}
\title{Alias for CRS.5b}
\usage{
ml4b(names = c("b", "c", "d", "e", "f"), fixed = c(NA, NA, NA, NA, NA), ...)
}
\description{
\code{ml4b} is an alias for \code{\link{CRS.5b}}.
}
\seealso{
\code{\link{CRS.5b}}
`r lifecycle::badge("deprecated")` 
This function is deprecated as of version 3.3.0. Please use [CRS.5()] instead.
}
\keyword{models}
\keyword{nonlinear}
