% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedergreen.R
\name{ml4c}
\alias{ml4c}
\title{Alias for CRS.5c}
\usage{
ml4c(names = c("b", "c", "d", "e", "f"), fixed = c(NA, NA, NA, NA, NA), ...)
}
\description{
\code{ml4c} is an alias for \code{\link{CRS.5c}}.
}
\seealso{
\code{\link{CRS.5c}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated as of version 3.3.0. Please use \code{\link[=CRS.5]{CRS.5()}} instead.
}
\keyword{models}
\keyword{nonlinear}
