\name{nicotine}

\alias{nicotine}

\docType{data}

\title{nicotine}

\description{nicotine data.}

\usage{data(nicotine)}

\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{conc}}{a numeric vector}
    \item{\code{total}}{a numeric vector}
    \item{\code{num.dead}}{a numeric vector}
  }
}

\keyword{datasets}
