% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.drc.R
\name{predict.drc}
\alias{predict.drc}
\title{Prediction}
\usage{
\method{predict}{drc}(
  object,
  newdata,
  se.fit = FALSE,
  interval = c("none", "confidence", "prediction", "ssd"),
  level = 0.95,
  na.action = na.pass,
  od = FALSE,
  vcov. = vcov,
  ssdSEfct = NULL,
  constrain = TRUE,
  checkND = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class 'drc'.}

\item{newdata}{an optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{se.fit}{logical. If TRUE standard errors are required.}

\item{interval}{character string. Type of interval calculation:
\code{"none"}, \code{"confidence"}, \code{"prediction"}, or \code{"ssd"}.}

\item{level}{tolerance/confidence level.}

\item{na.action}{function determining what should be done with missing values
in \code{newdata}. The default is to predict \code{NA}.}

\item{od}{logical. If TRUE adjustment for over-dispersion is used.}

\item{vcov.}{function providing the variance-covariance matrix.
\code{\link{vcov}} is the default, but \code{sandwich} is also an option
(for obtaining robust standard errors).}

\item{ssdSEfct}{specifies the function for interpolating standard errors
between observed standard errors. The default is linear interpolation on
log-log scale (back-transformed).}

\item{constrain}{logical. If TRUE (default) predicted values are truncated
within meaningful limits, i.e., 0 and, possibly, 1.}

\item{checkND}{logical indicating whether or not names in \code{newdata}
data frame match the names in the original data frame used for fitting
the model. Default is TRUE.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix with as many rows as there are dose values provided in
\code{newdata} or in the original dataset (in case \code{newdata} is not
specified) and, at most, 4 columns containing fitted values, standard
errors, lower and upper limits of confidence/prediction intervals.
}
\description{
Predicted values for models of class 'drc'.
}
\examples{
## Fitting a model
spinach.model1 <- drm(SLOPE~DOSE, CURVE, data = spinach, fct = LL.4())

## Predicting values at dose=2 (with standard errors)
predict(spinach.model1, data.frame(dose=2, CURVE=c("1", "2", "3")), se.fit = TRUE)

## Getting confidence intervals
predict(spinach.model1, data.frame(dose=2, CURVE=c("1", "2", "3")),
interval = "confidence")

## Getting prediction intervals
predict(spinach.model1, data.frame(dose=2, CURVE=c("1", "2", "3")),
interval = "prediction")

}
\seealso{
For details see the help page for \code{\link{predict.lm}}.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
