\name{red.fescue}

\alias{red.fescue}

\docType{data}

\title{Red fescue}

\description{red.fescue data.}

\usage{data(red.fescue)}

\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{biomass}}{a numeric vector}
  }
}

\keyword{datasets}
