% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsq.R
\name{Rsq}
\alias{Rsq}
\title{R-squared for dose-response models}
\usage{
Rsq(object)
}
\arguments{
\item{object}{an object of class 'drc'.}
}
\value{
Invisibly returns a matrix of R-squared values. For single-curve models, a 1x1 matrix.
  For multi-curve models, includes per-curve values and a total R-squared.
}
\description{
Calculates and displays R-squared values for a fitted dose-response model. For models
with multiple curves, per-curve and total R-squared values are returned.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
