\name{ryegrass2}

\alias{ryegrass2}

\docType{data}

\title{Ryegrass}

\description{Data from a dose-response experiment with ryegrass (\emph{Lolium} sp.). Biomass was measured at different dose levels and at two time points.}

\usage{data(ryegrass2)}

\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{biomass}}{a numeric vector}
    \item{\code{day}}{a categorial vector}
  }
}

\keyword{datasets}
