\name{ryegrass2}

\alias{ryegrass2}

\docType{data}

\title{Ryegrass}

\description{ryegrass2 data.}

\usage{data(ryegrass2)}

\format{
  A data frame with 27 observations on the following 3 variables.
  \describe{
    \item{\code{dose}}{a numeric vector}
    \item{\code{biomass}}{a numeric vector}
    \item{\code{day}}{a categorial vector}
  }
}

\keyword{datasets}
