% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdrc.R
\name{searchdrc}
\alias{searchdrc}
\title{Searching through a range of initial parameter values to obtain convergence}
\usage{
searchdrc(object, which, range, len = 50)
}
\arguments{
\item{object}{an object of class 'drc'. The object can be from a model that could not be fitted.}

\item{which}{a character string containing the parameter name.}

\item{range}{a numeric vector of length 2 specifying the interval endpoints for the range.}

\item{len}{numeric. The number of points in the interval.}
}
\value{
An object of class 'drc'.
}
\description{
\code{searchdrc} provides a facility for searching through a range of parameter values
(one-dimensional) in order to obtain convergence of the estimation procedure.
}
\details{
The function goes through the range with increments such that in total at most \code{len}
sets of parameter values are used as initial values for the estimation procedure. You would
need to identify the parameter which is most likely to cause problems for the estimation
procedure.
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
