\name{TCDD}

\alias{TCDD}

\docType{data}

\title{Liver tumor incidence}

\description{
  Liver tumor incidence in Sprague-Dawley rats exposed to the chemical like 2,3,7,8-tetrachlorodibenzo-pdioxin
(TCDD).
}

\usage{data(TCDD)}

\format{
  A data frame with 6 observations on the following 3 variables.
  \describe{
    \item{\code{conc}}{a numeric vector reporting the concentration of TCDD (ng/kg)}
    \item{\code{total}}{a numeric vector}
    \item{\code{incidence}}{a numeric vector}
  }
}

%\details{}

\source{
   R. Kociba, D. Keyes, J. Beyer, R. Carreon, C. Wade, D. Dittenber, R. Kalnins, L. Frauson,
C. Park, S. Barnard, R. Hummel, and C. Humiston (1978). Results of a two-year chronic toxicity
and oncogenicity study of 2,3,7,8-tetrachlorodibenzo-p-dioxin in rats. Toxicology and
Applied Pharmacology, \bold{46(2)}:279--303.
}

%\references{}



\keyword{datasets}
