% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{ucedergreen}
\alias{ucedergreen}
\title{U-shaped Cedergreen-Ritz-Streibig model}
\usage{
ucedergreen(
  fixed = c(NA, NA, NA, NA, NA),
  names = c("b", "c", "d", "e", "f"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  alpha
)
}
\arguments{
\item{fixed}{numeric vector. Specifies which parameters are fixed and at what value they are fixed.
NAs for parameters that are not fixed.}

\item{names}{a vector of character strings giving the names of the parameters (should not contain ":").
The order of the parameters is: b, c, d, e, f.}

\item{method}{character string indicating the self starter function to use.}

\item{ssfct}{a self starter function to be used.}

\item{alpha}{numeric value between 0 and 1, reflecting the steepness of the hormesis peak.
This argument must be specified.}
}
\value{
A list containing the non-linear function, the self starter function
  and the parameter names.
}
\description{
\code{ucedergreen} provides a very general way of specifying the Cedergreen-Ritz-Streibig
modified log-logistic model for describing u-shaped hormesis, under various constraints on the parameters.
}
\details{
The u-shaped model is given by the expression
\deqn{f(x) = c + d - \frac{d-c+f \exp(-1/x^{\alpha})}{1+\exp(b(\log(x)-\log(e)))}}
}
\references{
Cedergreen, N. and Ritz, C. and Streibig, J. C. (2005)
Improved empirical models describing hormesis,
\emph{Environmental Toxicology and Chemistry} \bold{24}, 3166--3172.
}
\seealso{
\code{\link{cedergreen}}, \code{\link{UCRS.4a}}, \code{\link{UCRS.5a}}, \code{\link{drm}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
