% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml3b}
\alias{uml3b}
\title{Alias for UCRS.4b}
\usage{
uml3b(names = c("b", "d", "e", "f"), ...)
}
\description{
\code{uml3b} is an alias for \code{\link{UCRS.4b}}.
}
\seealso{
\code{\link{UCRS.4b}}
}
\keyword{models}
\keyword{nonlinear}
