% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml3c}
\alias{uml3c}
\title{Alias for UCRS.4c}
\usage{
uml3c(names = c("b", "d", "e", "f"), ...)
}
\description{
\code{uml3c} is an alias for \code{\link{UCRS.4c}}.
}
\arguments{
\item{names}{a vector of character strings giving the names of the parameters.}

\item{...}{additional arguments passed to \code{\link{ucedergreen}}.}
}
\seealso{
\code{\link{UCRS.4c}}
}
\keyword{models}
\keyword{nonlinear}
