% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml4a}
\alias{uml4a}
\title{Alias for UCRS.5a}
\usage{
uml4a(names = c("b", "c", "d", "e", "f"), ...)
}
\description{
\code{uml4a} is an alias for \code{\link{UCRS.5a}}.
}
\seealso{
\code{\link{UCRS.5a}}
}
\keyword{models}
\keyword{nonlinear}
