% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucedergreen.R
\name{uml4b}
\alias{uml4b}
\title{Alias for UCRS.5b}
\usage{
uml4b(names = c("b", "c", "d", "e", "f"), ...)
}
\description{
\code{uml4b} is an alias for \code{\link{UCRS.5b}}.
}
\seealso{
\code{\link{UCRS.5b}}
}
\keyword{models}
\keyword{nonlinear}
