% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1.r
\name{W1.2}
\alias{W1.2}
\alias{w2}
\title{Two-parameter Weibull type 1 model}
\usage{
W1.2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)

w2(upper = 1, fixed = c(NA, NA), names = c("b", "e"), ...)
}
\arguments{
\item{upper}{numeric value giving the fixed upper limit. The default is 1.}

\item{fixed}{numeric vector of length 2. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that are not fixed.}

\item{names}{character vector of length 2 giving the names of the
parameters. The default is \code{c("b", "e")}.}

\item{\dots}{additional arguments passed to \code{\link{weibull1}}.}
}
\value{
A list of class \code{Weibull-1} containing the nonlinear function,
  self starter function, and parameter names.
}
\description{
A two-parameter Weibull type 1 model with the lower limit fixed at 0
and the upper limit fixed at a specified value (default 1).
}
\details{
The model is given by the expression
\deqn{f(x) = upper \exp(-\exp(b(\log(x) - \log(e))))}

This is mostly used for binomial/quantal responses.
}
\examples{
earthworms.m1 <- drm(number/total ~ dose, weights = total,
  data = earthworms, fct = W1.2(), type = "binomial")

}
\seealso{
\code{\link{weibull1}}, \code{\link{W1.3}}, \code{\link{W1.4}}
}
\keyword{models}
\keyword{nonlinear}
