% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull1.r
\name{W1.3}
\alias{W1.3}
\alias{w3}
\title{Three-parameter Weibull type 1 model}
\usage{
W1.3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)

w3(fixed = c(NA, NA, NA), names = c("b", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that are not fixed.}

\item{names}{character vector of length 3 giving the names of the
parameters. The default is \code{c("b", "d", "e")}.}

\item{\dots}{additional arguments passed to \code{\link{weibull1}}.}
}
\value{
A list of class \code{Weibull-1} containing the nonlinear function,
  self starter function, and parameter names.
}
\description{
A three-parameter Weibull type 1 model with the lower limit fixed at 0.
}
\details{
The model is given by the expression
\deqn{f(x) = d \exp(-\exp(b(\log(x) - \log(e))))}

This is a special case of the four-parameter Weibull type 1 model
where the lower limit is fixed at 0.
}
\examples{
ryegrass.m1 <- drm(rootl ~ conc, data = ryegrass, fct = W1.3())

}
\seealso{
\code{\link{weibull1}}, \code{\link{W1.2}}, \code{\link{W1.4}}
}
\keyword{models}
\keyword{nonlinear}
