% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.R
\name{W2.4}
\alias{W2.4}
\title{Four-parameter Weibull (type 2) model}
\usage{
W2.4(fixed = c(NA, NA, NA, NA), names = c("b", "c", "d", "e"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 4, specifying fixed parameters (use \code{NA} for
parameters that should be estimated).}

\item{names}{character vector of length 4 giving the names of the parameters
(default \code{c("b", "c", "d", "e")}).}

\item{...}{additional arguments passed to \code{\link{weibull2}}.}
}
\value{
A list of class \code{"Weibull-2"} as returned by \code{\link{weibull2}}.
}
\description{
A four-parameter Weibull type 2 model. The model is given by the equation
\deqn{f(x) = c + (d - c)(1 - \exp(-\exp(b(\log(x) - \log(e)))))}
}
\examples{
terbuthylazin.m1 <- drm(rgr ~ dose, data = terbuthylazin, fct = W2.4())

}
\seealso{
\code{\link{weibull2}}, \code{\link{W2.2}}, \code{\link{W2.3}}
}
\keyword{models}
\keyword{nonlinear}
