% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2x.R
\name{W2x.3}
\alias{W2x.3}
\title{Three-parameter Weibull type 2 model with lag time}
\usage{
W2x.3(fixed = c(NA, NA, NA), names = c("d", "e", "t0"), ...)
}
\arguments{
\item{fixed}{numeric vector of length 3. Specifies which parameters are
fixed and at what value. Use \code{NA} for parameters that should be
estimated (default is \code{c(NA, NA, NA)}).}

\item{names}{character vector of length 3 giving the names of the
parameters (default is \code{c("d", "e", "t0")}).}

\item{\dots}{additional arguments passed to \code{\link{weibull2x}}.}
}
\value{
A list of class \code{"Weibull-2"} containing the nonlinear
function, self starter function, and parameter names.
}
\description{
A three-parameter Weibull type 2 model with lag time, where \code{b} is
fixed at 1 and \code{c} is fixed at 0. This is a convenience wrapper
around \code{\link{weibull2x}}.
}
\examples{
spinach.m1 <- drm(SLOPE ~ DOSE, data = spinach, fct = W2x.3())
summary(spinach.m1)

}
\seealso{
\code{\link{weibull2x}}, \code{\link{W2x.4}}, \code{\link{W2.3}}
}
\keyword{models}
\keyword{nonlinear}
