% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull2.r
\name{weibull2}
\alias{weibull2}
\title{The four-parameter Weibull (type 2) model}
\usage{
weibull2(
  fixed = c(NA, NA, NA, NA),
  names = c("b", "c", "d", "e"),
  method = c("1", "2", "3", "4"),
  ssfct = NULL,
  fctName,
  fctText
)
}
\arguments{
\item{fixed}{numeric vector of length 4, specifying fixed parameters (use \code{NA} for
parameters that should be estimated).}

\item{names}{character vector of length 4 giving the names of the parameters
(default \code{c("b", "c", "d", "e")}).}

\item{method}{character string indicating the self starter method to use.
One of \code{"1"}, \code{"2"}, \code{"3"}, or \code{"4"}.}

\item{ssfct}{a self starter function. If \code{NULL} (default), a built-in
self starter is used based on \code{method}.}

\item{fctName}{optional character string used internally for the function name.}

\item{fctText}{optional character string used internally for the function description.}
}
\value{
A list containing the nonlinear function, self starter function,
and parameter names. The list has class \code{"Weibull-2"}.
}
\description{
Provides a general framework for the four-parameter Weibull type 2 model
given by the equation
\deqn{f(x) = c + (d - c)(1 - \exp(-\exp(b(\log(x) - \log(e)))))}
}
\references{
Seber, G. A. F. and Wild, C. J. (1989)
\emph{Nonlinear Regression}, New York: Wiley \& Sons (pp. 338--339).
}
\seealso{
\code{\link{weibull1}}, \code{\link{W2.2}}, \code{\link{W2.3}},
\code{\link{W2.4}}
}
\author{
Christian Ritz
}
\keyword{models}
\keyword{nonlinear}
